### 2020-05-04 #24 @mluerig
Phenopype: a phenotyping pipeline for Python
Submitting Author: Name (@mluerig)
Package Name: phenopype
One-Line Description of Package: a phenotyping pipeline for Python
Repository Link:  https://github.com/mluerig/phenopype
Version submitted:  1.0.5
Version accepted: 2.0.1

Edit: Bumped from 1.0.4 to 1.0.5 since submission.

---

Phenopype is a high throughput phenotyping pipeline for Python to support biologists in extracting high dimensional phenotypic data from digital images. The program provides intuitive, high level computer vision functions for image preprocessing, segmentation, and feature extraction. Users can assemble their own function-stacks that can be stored in the human-readable `yaml`-format along with raw data and results, facilitating high throughput and full data reproducibility. Phenopype can be run from Python or from a Python Integrated Development Environment (IDE), like Spyder. Phenopype is designed to provide robust image analysis workflows that can be implemented with little or no Python experience.

- Please indicate which [category or categories](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html) this package falls under:

\* Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [this section](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html#notes-on-categories) of our guidebook.

Phenopype is designed to extract phenotypic data (https://en.wikipedia.org/wiki/Phenotype) of plants, animals, and other organisms from images and videos. By processing images in Python, through reproducible code and human readable configuration files, phenotyping becomes higher throughput and more reproducible than established GUI programs like "ImageJ".

Phenopype is intended for ecologists and evolutionary biologists that work with phenotypic data. Phenotypic data are an essential component of ecological and evolutionary research (https://www.nature.com/articles/nrg2897).

Only low level computer vision packages like OpenCV or scikit-image are out there that require a lot of configuring and a relatively deep understanding of computer vision and Python in general. Phenopype offers high level functions so that users can focus on the relevant analytic parts of image analysis.

    - Full functionality is currently given for Windows and Linux builds. Under macOS, the high throughput workflow does not work due to two conflicting dependencies (watchdog and OpenCV)
    - Although phenopype attempts to minimize user interaction with images (e.g. setting ROIs or landmarks), this is a still a key component. I have done what I can to mimic user input for testing and CI, but I could not find a way to fake mouse movement and clicking, which is why the testing suite does not cover some of the lowlevel functions. Therefore, coverage is about 10% lower than what it could be with these interactions.
    - This is the presubmission inquiry: https://github.com/pyOpenSci/software-review/issues/23

- [X] has an [OSI approved license](https://opensource.org/licenses)

- [ ] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements). Be aware that completing the pyOpenSci review process **does not** guarantee acceptance to JOSS. Be sure to read their submission requirements (linked above) if you are interested in submitting to JOSS.
- [ ] The package is not a "minor utility" as defined by JOSS's [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements): "Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable." pyOpenSci welcomes these packages under "Data Retrieval", but JOSS has slightly different criteria.

- [X] I agree to abide by [pyOpenSci's Code of Conduct](https://www.pyopensci.org/dev_guide/peer_review/coc.html) during the review process and in maintaining my package should it be accepted.

**P.S.** *Have feedback/comments about our review process? Leave a comment [here](https://github.com/pyOpenSci/governance/issues/8)*

[Editor and review templates can be found here](https://www.pyopensci.org/contributing-guide/appendices/templates.html)
