+++ docs/_sources/example_2_landmarks_stickleback.ipynb.txt
@@ -1,0 +1,981 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Example 2: Stickleback morphometrics - landmarks\n",
+    "\n",
+    "Functional morphology of organisms is often measured by placing landmarks at specific points that show structural, functional or developmental significance. In this example phenopype is used to place morphometric landmarks across the anterior half of a stickleback (*Gasterosteus aculeatus*) stained with alizarin red.\n",
+    "\n",
+    "First we place landmarks in low throughput mode to learn how the landmark-function works, then we look at a high throughput landkmark example with a project directory and a global scale. "
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "<div class=\"row; text-align: left\">\n",
+    "    \n",
+    "<div class=\"col-md-6\">\n",
+    "    \n",
+    "![Before](_assets/ex2_before.jpg)\n",
+    "    \n",
+    "**Input** - Stained threespine stickleback, photographed in a glycerol bath from a camera stand\n",
+    "</div>\n",
+    "<div class=\"col-md-6\">\n",
+    "\n",
+    "![After](_assets/ex2_after.jpg)\n",
+    "    \n",
+    "**Results** - 22 landmarks are placed using the `landmark` tool from the `phenopype.measurements` module\n",
+    "</div>\n",
+    "</div>\n"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Low throughput"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import phenopype as pp\n",
+    "\n",
+    "filepath = r\"images/stickle1.jpg\"\n",
+    "\n",
+    "ct = pp.load_image(filepath, cont=True) ## load image as container"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "- setting landmarks\n"
+     ]
+    }
+   ],
+   "source": [
+    "## just place a few test-landmarks\n",
+    "\n",
+    "pp.measurement.landmarks(ct, \n",
+    "                         point_size=15, \n",
+    "                         point_colour=\"green\", \n",
+    "                         label_size=1, overwrite=True) \n",
+    "\n",
+    "## sets landmarks. the landmarks get stored inside the container. if you have already placed landmarks\n",
+    "## to the same container and set \"overwrite=False\", you wont be able to place them again until you set\n",
+    "## the overwrite flag to \"True\""
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "To create an image with the selected points, we first need te select the background image (or \"canvas\") with `select_canvas` and then call the `draw_landmarks` function. These explicit steps are necessary when using the low throughput, but not the high throughput routine (see [Tutorial 2](tutorial_2_phenopype_workflow.ipynb)); likewise the `save_landmarks` function."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "- raw image\n"
+     ]
+    },
+    {
+     "data": {
[…]

[…]
