### 2020-05-18 #24 @lwasser
awesome. thank you @mluerig !!! if i don't hear back on twitter i'll reach out to some folks directly.
