### 2020-05-19 #24 @lwasser
ok @mluerig i found another person who might be interested as well. @sdonoughe and @agporto are you still interested in reviewing this package? please let me know... i'm still looking for someone to serve as a guest editor. I also have a ping from Ben Cook who i don't have a github handle for yet who may be available for review or editor role!!  All -- please let me know if you are available!! i will watch this thread for responses. And i'll keep looking for a guest editor in case everyone wants to review! thank you!!
