### 2020-05-19 #24 @jbencook
Hey All! I'm new here, but I'm available to review if it's helpful. I lead the Applied Machine Learning team at Hudl where we build computer vision services for sports video. I have some background in scientific computing and I'd love to help if I can.
