### 2020-05-19 #24 @jbencook
I'm also happy to serve as editor instead if that's where the need is
