### 2020-05-19 #24 @lwasser
@jbencook ok awesome!! Let's see if our other two colleagues respond!! I can absolutely walk you through serving as an editor. thank you again for being so responsive and willing to help! and welcome to pyopensci!!
