### 2020-05-19 #24 @lwasser
perfect. thank you @agporto ! i'll followup with instructions once.i see if we have a third person or not!! :) so appreciative of you all being willing to work with pyopensci!!
