### 2020-05-19 #24 @sdonoughe
I don't have experience reviewing packages, but I fall squarely into the
intended audience/user-set for the package, so I am happy to participate.

On Tue, May 19, 2020 at 2:16 PM Leah Wasser <notifications@github.com>
wrote:

> perfect. thank you @agporto <https://github.com/agporto> ! i'll followup
> with instructions once.i see if we have a third person or not!! :) so
> appreciative of you all being willing to work with pyopensci!!
>
