### 2020-05-19 #24 @lwasser
👋 there @sdonoughe sure... tell me a bit more. do you have enough programming background that you feel comfortable reviewing with another person as a team? You could also test out the functionality and documentation whereas @agporto could focus on the code / back end. please just let me know what works best for you and if you have questions you can also email me directly at leah.wasser@colorado.edu !! thank you for being willing to participate.
