### 2020-05-22 #24 @lwasser
ok everyone... i think we are in good shape here. @jbencook if you are still open to serving as a guest editor, that would be great. I will be out next week. While i am out, @jlpalomino has agreed to answer any questions that you have about the review process. We just updated our [contributing guide](https://www.pyopensci.org/contributing-guide/intro.html) so i'm hoping better instructions will be available to you ONCE we figure out why github actions is acting funny when trying to deploy the updates!

@agporto your review is much appreciated! if you can focus more on the technical end of things - code, tests etc that would be great! @sdonoughe if you can please focus on the usability and docs + applications of the package for your needs / and general use cases in the community that would be excellent. I am working on also finding another person to support this review so you can ask questions as you go!! that can be me once i'm back on june 1.

@mluerig i just also want to confirm that you are NOT interested in submitting this to JOSS. i noticed that was not checked. This process requires that you write a short paper. but JOSS accepts our review once your package is approved so it is not a challenging additional step.

@jbencook the next steps are for you to do the following following [our guidelines](https://www.pyopensci.org/contributing-guide/submissions/editor_guide.html)
- [ ] Add your editor checks using this template:  https://221-165885836-gh.circle-artifacts.com/0/html/appendices/templates.html
- [ ]  i've just updated the issue template at the top to list you as an editor and the other two reviewers.
- [ ]  please provide the reviewers with a review deadline and direct them to our review template - https://www.pyopensci.org/contributing-guide/appendices/templates.html#review-template . We normally suggest 3 weeks . given i will be out next week and it is a holiday weekend here in the states - we could allow for a bit more time if that is useful to everyone
- [ ] check in with the reviewers 2 weeks in to ensure things are going smoothly!
- [ ] watch the reviewers interactive with the package author and see if there are any questions.

All please look at our documentation here:   https://www.pyopensci.org/contributing-guide/ we are working on improving it so any feedback is much appreciated!!

Please all get in touch with any questions or concerns (knowing i'll just be offline for one week if you need feedback from me!). Thank you ALL for supporting pyopensci!!
