### 2020-05-22 #24 @jlpalomino
Thanks @lwasser for providing @jbencook with the next steps. Happy to answer any questions about the editor process, so feel free to reach out!

While we wait for the updates to the guideline docs to render, we can see the [latest version here](https://github.com/pyOpenSci/contributing-guide/blob/master/open-source-software-submissions/editors-guide.md).
