### 2020-05-23 #24 @jbencook
## Editor checks:

- [ ] **License:** The package has an OSI accepted license
- [ ] ~~**Archive** (JOSS only, may be post-review): The repository DOI resolves correctly~~
- [ ] ~~**Version** (JOSS only, may be post-review): Does the release version given match the GitHub release (v1.0.0)?~~

---

## Editor comments

---

Reviewers: @agporto, @sdonoughe
Due date: **2020-06-19**
