### 2020-05-23 #24 @jbencook
Thanks @lwasser and good to meet you @jlpalomino!

I just added the editor checks. @agporto and @sdonoughe does 19 June work for the review deadline? I'll start watching the repo and work on my checks.

Also, @mluerig I noticed the repo has an [LGPL license](https://github.com/mluerig/phenopype/blob/895d243b6b119f67d96baee6df5d3cd019c02984/LICENSE.txt), but the README says [MIT](https://github.com/mluerig/phenopype/blame/0e94c3f775abeea285ec62e1e13b07c8c1f777a3/README.md#L4). Looks like both are approved OSI licenses, but they should probably match.

Looking forward to working with you all!
