### 2020-05-24 #24 @mluerig
@lwasser @jbencook @jlpalomino yupp correct I am NOT planning to submit to JOSS, but to Methods in Ecology and Evolution. The manuscript is already written and can be made available to anyone involved in the review process, if that is needed and helpful
