### 2020-05-24 #24 @mluerig
@jbencook I changed it to LGPL in the readme
