### 2020-06-03 #24 @lwasser
hi all!! @jbencook just a note that i have slowed the process down here as we are looking for someone to support the second review. With pyopensci we want to mentor folks through reviews if it is their first. i however was out last week and i hadn't found someone prior to being out. So i'm on the hunt again and we will get back to you with a timeline for the second review! if you know of any people who have experience with code reviews that could support a second review please do say the word.
