+++ phenopype/core/preprocessing.py
@@ -1,0 +1,785 @@
+#%% modules
+import cv2, copy
+import numpy as np
+import pandas as pd
+
+from math import sqrt
+import numpy.ma as ma
+
+from phenopype.settings import colours
+from phenopype.utils import load_image_data
+from phenopype.utils_lowlevel import _image_viewer, _equalize_histogram, _contours_arr_tup
+from phenopype.utils_lowlevel import _auto_text_size, _auto_text_width
+
+#%% functions
+
+
+def create_mask(
+    obj_input,
+    df_image_data=None,
+    include=True,
+    label="mask1",
+    overwrite=False,
+    tool="rectangle",
+    **kwargs
+):
+    """
+    Draw rectangle or polygon mask onto image by clicking and dragging the 
+    cursor over the image. One mask can contain multiple sets of coordinates, 
+    i.e. multiple and not overallping masks. For rectangle mask, finish with 
+    ENTER. For polygons, finish current polygon with CTRL, and then with ENTER.
+    
+    Parameters
+    ----------
+    obj_input : array or container
+        input object
+    df_image_data : DataFrame, optional
+        an existing DataFrame containing image metadata, will be added to mask
+        output DataFrame
+    include: bool, optional
+        determine whether resulting mask is to include or exclude objects within
+    label: str, optinal
+        assigns a label to the mask
+    overwrite: bool, optional
+        if working using a container, or from a phenopype project directory, 
+        should existing masks with the same label be overwritten
+    tool: {"rectangle", "polygon"} str, optional
+        select tool by which mask is drawn
+
+    Returns
+    -------
+    df_masks: DataFrame or container
+        contains mask coordiantes
+    """
+
+    ## kwargs
+    flag_overwrite = overwrite
+    test_params = kwargs.get("test_params", {})
+
+    ## load image
+    df_masks = None
+    if obj_input.__class__.__name__ == "ndarray":
+        image = obj_input
+        if df_image_data.__class__.__name__ == "NoneType":
+            df_image_data = pd.DataFrame(
+                {"filename": "unknown"}, index=[0]
+            )  ## that may not be necessary
+    elif obj_input.__class__.__name__ in ["container", "motion_tracker"]:
+        image = copy.deepcopy(obj_input.image)
+        if hasattr(obj_input, "df_image_data"):
+            df_image_data = copy.deepcopy(obj_input.df_image_data)
+        if hasattr(obj_input, "df_masks"):
+            df_masks = copy.deepcopy(obj_input.df_masks)
+    else:
+        print("wrong input format.")
+        return
+
+    ## check if exists
+    while True:
+        if not df_masks.__class__.__name__ == "NoneType" and flag_overwrite == False:
+            df_masks = df_masks[
+                df_masks.columns.intersection(["mask", "include", "coords"])
+            ]
+            if label in df_masks["mask"].values:
+                print("- mask with label " + label + " already created (overwrite=False)")
+                break
+        elif not df_masks.__class__.__name__ == "NoneType" and flag_overwrite == True:
+            df_masks = df_masks[
+                df_masks.columns.intersection(["mask", "include", "coords"])
+            ]
+            if label in df_masks["mask"].values:
+                df_masks = df_masks.drop(df_masks[df_masks["mask"] == label].index)
+                print("- create mask (overwriting)")
+                pass
+        elif df_masks.__class__.__name__ == "NoneType":
+            print("- create mask")
+            df_masks = pd.DataFrame(columns=["mask", "include", "coords"])
+            pass
+
+        ## method
+        out = _image_viewer(image, mode="interactive", tool=tool, previous=test_params)
+
+        ## abort
+        if not out.done:
+            if obj_input.__class__.__name__ == "ndarray":
+                print("terminated mask creation")
+                return
+            elif obj_input.__class__.__name__ == "container":
[…]
