+++ phenopype/core/measurement.py
@@ -1,0 +1,690 @@
+#%% modules
+import cv2, copy
+import math
+import numpy as np
+import numpy.ma as ma
+import pandas as pd
+
+from phenopype.utils_lowlevel import (
+    _image_viewer,
+    _auto_line_width,
+    _auto_point_size,
+    _auto_text_width,
+    _auto_text_size,
+)
+
+#%% methods
+
+
+def landmarks(
+    obj_input,
+    df_image_data=None,
+    overwrite=False,
+    point_colour="green",
+    point_size="auto",
+    label_colour="black",
+    label_size="auto",
+    label_width="auto",
+    **kwargs
+):
+    """
+    Place landmarks.
+    
+    Parameters
+    ----------
+    obj_input : array or container
+        input object
+    df_image_data : DataFrame, optional
+        an existing DataFrame containing image metadata, will be added to landmark
+        output DataFrame
+    overwrite: bool, optional
+        if working using a container, or from a phenopype project directory, 
+        should existing landmarks be overwritten
+    point_colour: {"green", "red", "blue", "black", "white"} str, optional
+        landmark point colour
+    point_size: int, optional
+        landmark point size in pixels
+    label_colour : {"black", "white", "green", "red", "blue"} str, optional
+        landmark label colour.
+    label_size: int, optional
+        landmark label font size (scaled to image)
+    label_width: int, optional
+        landmark label font width  (scaled to image)
+
+    Returns
+    -------
+    df_masks: DataFrame or container
+        contains landmark coordiantes
+    """
+
+    ## kwargs
+    flag_overwrite = overwrite
+    test_params = kwargs.get("test_params", {})
+
+    ## load image
+    df_landmarks = None
+    if obj_input.__class__.__name__ == "ndarray":
+        image = obj_input
+        if df_image_data.__class__.__name__ == "NoneType":
+            df_image_data = pd.DataFrame({"filename": "unknown"}, index=[0])
+    elif obj_input.__class__.__name__ == "container":
+        image = copy.deepcopy(obj_input.image_copy)
+        df_image_data = obj_input.df_image_data
+        if hasattr(obj_input, "df_landmarks"):
+            df_landmarks = obj_input.df_landmarks
+    else:
+        print("wrong input format.")
+        return
+
+    ## more kwargs
+    if point_size == "auto":
+        point_size = _auto_point_size(image)
+    if label_size == "auto":
+        label_size = _auto_text_size(image)
+    if label_width == "auto":
+        label_width = _auto_text_width(image)
+
+    while True:
+        ## check if exists
+        if not df_landmarks.__class__.__name__ == "NoneType" and flag_overwrite == False:
+            df_landmarks = df_landmarks[
+                df_landmarks.columns.intersection(["landmark", "x", "y"])
+            ]
+            print("- landmarks already set (overwrite=False)")
+            break
+        elif not df_landmarks.__class__.__name__ == "NoneType" and flag_overwrite == True:
+            print("- setting landmarks (overwriting)")
+            pass
+        elif df_landmarks.__class__.__name__ == "NoneType":
+            print("- setting landmarks")
+            pass
+
+        ## set landmarks
+        out = _image_viewer(
+            image,
+            tool="landmarks",
+            point_size=point_size,
+            point_colour=point_colour,
+            label_size=label_size,
+            label_width=label_width,
+            label_colour=label_colour,
+            previous=test_params,
+        )
+        coords = out.points
+
+        ## abort
+        if not out.done:
+            if obj_input.__class__.__name__ == "ndarray":
+                print("terminated polyline creation")
+                return
+            elif obj_input.__class__.__name__ == "container":
[…]

[…]
