### 2020-06-16 #24 @jbencook
I've completed the editor checks - I think now we're just waiting on the reviews. How are things coming @agporto and @sdonoughe? Any questions so far?
