### 2020-06-19 #24 @sdonoughe
Hi folks, I've also been delayed, but aiming to conclude by the end of next
week.

On Fri, Jun 19, 2020 at 4:24 PM Arthur Porto <notifications@github.com>
wrote:

> Hi everyone. I am going to need a few more days to finish the review (I
> expect Tuesday next week). I apologize for the delay. The source code of
> the package is more extensive than I initially thought.
>
