### 2020-06-25 mluerig/phenopype#7 @agporto
Tutorial 2 - AttributeError: 'NoneType' object has no attribute 'shape'
When running Tutorial 2, under the Low throughput workflow, I am getting the following error:

```
<ipython-input-5-3766e430d9d3> in <module>
     15                            kernel_size=3, iterations=3) # 5
     16 pp.segmentation.find_contours(container, retrieval="ext", min_area=150) # 6
---> 17 pp.visualization.draw_contours(container) # 6
     18 pp.export.save_contours(container, dirpath = r"../_temp/output")
     19 pp.show_image(container.canvas)

~/phenopype/core/visualization.py in draw_contours(obj_input, df_contours, offset_coords, label, fill, mark_holes, level, line_colour, label_size, label_colour, line_width, label_width, skeleton, watershed, bounding_box, bounding_box_ext)
    190     ## more kwargs
    191     if line_width == "auto":
--> 192         line_width = _auto_line_width(image)
    193     if label_size == "auto":
    194         label_size = _auto_text_size(image)

~/phenopype/utils_lowlevel.py in _auto_line_width(image, **kwargs)
    642 def _auto_line_width(image, **kwargs):
    643     factor = kwargs.get("factor", 0.001)
--> 644     image_height, image_width = image.shape[0:2]
    645     image_diagonal = (image_height + image_width) / 2
    646     line_width = max(int(factor * image_diagonal), 1)

AttributeError: 'NoneType' object has no attribute 'shape'
```

Based on an inspection of the source code, it seems that nothing is getting assigned to `container.canvas`. My attempts to trace it back haven't yielded anything useful so far. I am just creating this issue to reference in my review.
