### 2020-06-26 mluerig/phenopype#8 @agporto
Tutorial 5 - Unexpected Keyword
When running Tutorial 5, under the polylines demo, I am getting the following error:
```
<ipython-input-6-a3de9e316ea9> in <module>
      2                                  line_width=2,
      3                                  line_colour="green")
----> 4 canvas = pp.visualization.draw_polylines(canvas, df_polylines=df_lines,
      5                                          line_width=2,
      6                                          line_colour="green")

TypeError: draw_polylines() got an unexpected keyword argument 'df_polylines'
```
Based on an inspection of the source code, the draw_polylines function definition does not contain the keyword in question, so my assumption is that some recent changes to the code have removed the need to explicitly define it and the tutorial has not been updated to reflect those changes.
