### 2020-06-26 #24 @agporto
- [X] As the reviewer I confirm that there are no conflicts of interest for me to review this work

- [X] **Function Documentation:** for all user-facing functions
- [X] **Examples** for all user-facing functions

      I could not find any reference to community guidelines

- [X] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [X] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [X] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [X] Any additional setup required (authentication tokens, etc)
- [ ] Brief demonstration usage

        The readme itself does not contain usage demonstration, but it links to it.

- [X] Direction to more detailed documentation (e.g. your documentation files or website).
- [X] If applicable, how the package compares to other similar packages and/or how it relates to other packages

        The package is still unpublished, so it does not contain citation information.

Package structure should follow general community best-practices. In general please consider:

- [X] The documentation is easy to find and understand

      See review below for a few exceptions.

      While not all functions contain automated tests, the exceptions are related to difficulties in replicating human input.

- [X] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

---

I have now reviewed the package entitled “Phenopype: a phenotyping pipeline for Python”. This package represents a tremendous amount of work, especially considering that it has a single author/contributor.
I strongly support its publication at pyOpenSci. The python code is clearly written and easy to follow. I believe it will be a critically important package for biologists interested in computer vision, because it is not only comprehensive but also offers a lot of flexibility in terms of usage. I have only a few comments with regards to the usage and code.

##### General suggestions

The documentation is quite extensive and well organized, but I think one thing that would help new users is moving the Tutorial on GUI interactions (currently, Tutorial 5) to the beginning. This tutorial explains in more detail how to interact with OpenCV's High GUI module. In my experience with the current package, this module was **by far** the hardest one to interact with. It is somewhat unstable, crashing the jupyter kernel with some frequency, and its keystrokes are not intuitive. Given that this module is important in mediating user interactions with Phenopype, I think it would be useful to have a tutorial on it front and center. It might prevent new users from getting overly frusfrated and ending up abandoning the package. In the long run (beyond this review), it might be worth considering other options for GUI interaction.

Another aspect that might need some attention is the standardization of the root directory. When going through the tutorials, perhaps it would be useful to explicitely set the root directory, since the root directory jumps around a little across the different tutorials (sometimes it starts on one place, sometimes on another)

##### Tutorial and Examples - Code Errors

When running Tutorial 2 and going through the 'Low throughput workflow' I get the following error:

```
~/phenopype/utils_lowlevel.py in _auto_line_width(image, **kwargs)
    642 def _auto_line_width(image, **kwargs):
    643     factor = kwargs.get("factor", 0.001)
--> 644     image_height, image_width = image.shape[0:2]
    645     image_diagonal = (image_height + image_width) / 2
    646     line_width = max(int(factor * image_diagonal), 1)

AttributeError: 'NoneType' object has no attribute 'shape'
```
Based on an inspection of the source code, it seems that, when first initializing a container and trying to draw a contour, no image is getting assigned to container.canvas. As a consequence, any attempt to extract the image parameters (height, width) from an empty object leads to the error above. It could be an easy fix, but my attempts at tracing it back have not yielded anything other than what is described above.

Similarly, when running Tutorial 5 and going through the polylines example, I get the following error:

```
<ipython-input-6-a3de9e316ea9> in <module>
      2                                  line_width=2,
      3                                  line_colour="green")
----> 4 canvas = pp.visualization.draw_polylines(canvas, df_polylines=df_lines,
      5                                          line_width=2,
      6                                          line_colour="green")

TypeError: draw_polylines() got an unexpected keyword argument 'df_polylines'
```

This one seems more straightforward, since the `draw_polylines` function definition does not contain the keyword in question. So, my assumption is that some recent changes to the code have removed the need to explicitly define `df_polylines` and the tutorial has not been updated to reflect those changes.

##### Code architecture - Suggestions

I have a suggestion in terms of code architecture that I do not expect the author to address in this review (but the author might find it useful in the long run). As of now, phenopype makes an explicit choice with regards to how to store data (both raw and processed data). Whether working with masks, polylines, landmarks, etc, phenopype will store that information (most of the time) in human-readable csv files or similar. It will also store different types of information (e.g., landmarks or masks) in different files using different functions. While this makes accessing the data for each specimen quite easy and in human-readable format, it also creates a lot of code duplication due to the need of having one function for each type of annotation. Given that this package will likely become a reference for computer vision in ecology and evolution, it will also set data standards for future work. Having followed the success of the COCO (https://cocodataset.org/#home) data format, I wonder if phenopype would benefit from having a similar architecture for data storage. By storing different types of annotations in a single JSON file, COCO has created an almost universal file format that makes it remarkably easy to train different machine learning algorithms using the same dataset (whether trying to predict landmarks or masks, for example). Again, I do not expect this change to occur, but I think it might be worth considering in the future.

##### Final remarks

In general, this is a well written and user friendly python package. My comments here are purely to provide the author with some (hopefully) useful feedback.

##### Other minor comments

- Some images in the image folder have 'JPG' (capitalized) extensions. These do not match some of the example notebooks (e.g. Example 2). These should probably be corrected, so as not to throw out a 'File not found' error simply due to capitalization.

- Hyperlink on Readme.md still links to MIT license.
