### 2020-07-01 #24 @sdonoughe
Hi again, the document and tutorial materials are extensive and impressive! As such it's taken me longer than anticipated to review all of the components. Will conclude soon.
