### 2020-07-01 #24 @mluerig
Hi all. I just realized that I probably should have provided this information at the beginning to make your lives a little easier: I wont be working for the next couple of weeks because of a long planned vacation. it's not yet clear for how long, but probably until early August. Sorry about not communicating my timeline earlier, I hope this is ok for everyone.
