### 2020-08-27 c105547
release 1.0.6
