### 2020-09-07 mluerig/phenopype#7 @mluerig
in the tutorial a line with `pp.visualization.select_canvas(container, canvas="raw")
` was missing, so no image was provided for canvas and `NoneType` returned

this is addressed in https://github.com/mluerig/phenopype/commit/c105547f1134b74aeda55b289a030c1f1c9e616b
