### 2020-09-07 mluerig/phenopype#8 @mluerig
indeed, a recent argument renaming (df_lines to df_polylines) was not reflected here

this is addressed in https://github.com/mluerig/phenopype/commit/c105547f1134b74aeda55b289a030c1f1c9e616b
