### 2020-10-15 #24 @mluerig
hey @lwasser - I have been in touch with @sdonoughe because he had some problems getting the program to work on a mac (I programmed and tested it only on windows and ubuntu). I think it's working now, and he is back to the actual review of documentation and tutorials. I am not sure how far along he is though.

In general, how are code-reviews typically addressed? I already fixed two bugs that @agporto spotted, but I will also streamline the tutorials according to his suggestions. Should I let everyone know when a new version is ready to be reviewed?

In that context a potentially complicating issue is that I continued development and released new versions since my submission to pyOpenSci, that have some new features. This was due to ongoing projects that needed these features. It's nothing fundamentally new or complex, I just extended the palette of functions that phenopype has, the basic principle and functionality is still the same. Should a second round of review just take a quick pass at the new features?
