### 2020-10-15 #24 @lwasser
hey there @mluerig this is a good question. it *sounds* like it might make sense for @sdonoughe to do another pass at using the package (if there is time - please let us know). if you have released new versions that have both fixes and enhancements i think for us to approve the package we will need to ensure that version has been reviewed. so let's see if @agporto and @sdonoughe have time to do another review on the current version. i think a quick pass to ensure things work as intended is sufficient. thank you for the speedy reply!
