### 2020-10-15 #24 @mluerig
ok great, but I would suggest @sdonoughe finishes the review on the current version. then I can address both his and @agporto's remaining comments together, and release another, post-review version

a second round of reviews on this version could then focus on the parts that received comments in the first round, and that are new as per the continued development. together with completed checks and CI on all new features, would you then be able to approve the package, given that there are no more comments?

hope that makes sense...
