### 2020-10-15 #24 @lwasser
@mluerig absolutely. whatever process that is the most efficient is the one we should proceed with! Our main concern at pyopensci is that whatever package version we stamp as approved does need to be reviewed in its entirety. The workflow that you propose above sounds reasonable if both reviewers have bandwidth!
