### 2020-11-02 #24 @sdonoughe
This sounds great to me. A quick update on my end. As @mluerig mentioned, I've been trying to get the package working on MacOS. As best I can tell it is an issue with system-specific differences with how OpenCV takes user input, which breaks the parts of phenopype's workflow. @mluerig has submitted a bug report about this, I believe, and so far there hasn't been a fix. I am now seeing if there is a trivial fix available by tweaking the user input steps in the phenopype code itself. I will work on this for another day or two, and if I can't get it working, I will just finish up my review on the rest of the package and on the tutorials.
