### 2021-02-04 #24 @sdonoughe
_Relevant disclosure:_ My written review of this package has taken a very long time for me to submit (~8 months from the time I was invited). Towards the end of that period, I contributed as a minor author to an as-yet-unpublished literature review manuscript on the topic of computer vision for biology research. This author of `phenopype` is one of several other co-authors on that manuscript. Since I began this review without being acquainted the author, and since I did the bulk of the code review before we began to interact as co-authors, I do not consider this to be a conflict of interest. But in the interest of full disclosure, I mention it here.

- [x] **Function Documentation:** for all user-facing functions
- [x] **Examples** for all user-facing functions
- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [x] Any additional setup required (authentication tokens, etc)
- [ ] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages

>_Reviewer notes on README:_
>
>**Brief demonstration usage:** There currently isn't a demonstration usage directly in the README, but there are, of course, lots of links to demonstrations. If a demo is needed here, perhaps the best would be have code that opens an image, takes user mouse clicks to select a region, and then does, say, a threshold operation on the contents of that region, then displays and saves the results.
>
>**How the package compares to similar packagese:** I think it would be useful to have an explanation for how this package relates to the most widely used Python image manipulation packages, and why a user might want to use `phenopype` _in combination_ with one or more of those packages. As for the graphical interface for taking user input interactively, the closest Python tool that I'm aware of is [napari](https://github.com/napari/napari), even though it is still in alpha. For users that do not have much coding experience, another option for them is to use [Fiji/ImageJ](https://imagej.net/Fiji), and then record a macro of their image processing steps. Here, `phenopype` offers functionality that is not present in any of these alternatives, so it would be useful to point that out too.
>
>**Citation information:** If/when this package is described in a journal article, then that citation can be added, but even in the meantime I can see biologists wanting to try `phenopype` out and potentially even using it in publications. So I'd definitely recommend having a "how to cite phenopype" section.

Package structure should follow general community best-practices. In general please consider:

- [x] The documentation is easy to find and understand

>_Reviewer notes on Usability:_ All the functions are well-documented in the code base itself, which is great. The API reference page is also quite helpful.

- [ ] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

>_Reviewer notes on Functionality:_ I was able to assess most of the functionality of package, but there were a few minor things that I was unable to check. The provided examples and the other most common use-cases I can imagine are not particularly computationally intensive; so it makes sense that there aren't any specific performance claims. I did not assess any automated tests, nor continuous integration, nor compliance with packaging guidelines (also, a sidenote to the Editor: that link in the Review template to packaging guidelines is broken)

#### For packages co-submitting to JOSS

---

_Note:_ My aim in this review was to focus on the scope, usability, documentation, tutorials, and user-facing functionality. I have not attempted to assess the clarity, organization, and performance of the code itself.

##### Aim/scope of the package
In many areas of biology search, a common challange is to take an image of one or more organisms, then measure some traits of those organisms. Such traits could include: the number of individual organisms; or a given organism's length, color spectrum, orientation, area of its spots, etc. After aquiring the raw images, this process entails many substeps:

- discarding faulty or misleading images
- determining how to usefully preprocess images with transformations in order to remove noise or highlight the traits of interest
- assessing whether those transformations introduce any artifacts
- getting interactive user input (such as mouse clicks) to select points or regions of interest
- measuring the traits of interest
- storing both the results and the sequence of image processing steps (along with any relevant parameters)
- applying a stored sequence of operations to new datasets

`phenopype` is intended to make all of these steps easier.

##### How does this package fit into an image analysis workflow?
One might wonder: why wouldn't a biologist just use `opencv` by itself? Or `scikit-image`, or any other Python library of image processing tools, for that matter? Answer: Because `phenopype` is aiming to augment, rather than replace, the utility of existing CV libraries for scientists measuring phenotypes. Put differently, `phenopype` is not a library of granular image processing functions. Instead, it is a set of wrappers around some of `opencv`'s most broadly used functions, combined with a simplified GUI for taking user input; these collectively act as a scientist's bookkeeper during an image processing project. In doing so, it fills a major void in the scientific Python toolkit. As far as I know there is no other Python tool that enables the user to create a pipeline of image-processing steps that incorporates manual input along the way. `phenopype` is also (to my knowledge) unique in defining a terse, yet legible format for recording and replicating the parameters of such a pipeline.

One side effect of the decision to rely on `opencv`, unfortunately, is that its image viewer functionality is currently buggy for me on MacOS. I include a truncated account of my efforts to get that part of the package to function correctly below under the **Installation** section.

There are some tradeoffs involved in structuring the package this way. For one, the set of image-manipulation functions is impressive, but it understandably covers only a small subset of the image-parsing tasks that are regularly encountered in image analysis work for biological data. This is not a knock against `phenotype` in the slightest. But a user that wanted to integrate additional image-processing functions into the middle of a workflow would need to invest some time to work out how `phenopype` works under the hood if they wanted to take advantage of the replicability tools that are the core of the package.

The upshot of this is that a user without much coding experience would probably need to restrict themselves to analyses that can be composed entirely of the functions that are folded into `phenopype` itself already, but as the tutorials and examples show, this includes a fairly wide range of potential analyses.

##### Installation
Notes on the _Installation_ page itself:
- It is not clear if it is important for `spyder` to be used with `phenopype`, or whether that is just a preference/recommendation of the author.
- It is ambiguous whether the user is recommended to install the updates and/or hotfixes.
- This is a weakly held opinion, but since the author thinks that installation into a new `conda` environment is a good idea (and I agree, by the way), I think that recommendation should come first, before the the line with `pip install phenopype`.
- There is a line "Mixing the two installers may break your python enviroment." I think this can be a little more precise, since earlier on the page there is a recommendation to use both in combination. At the time of this writing, my understanding is that the best practice is to make a `conda` environment, then use `conda` to add whatever additional packages are needed, and then do all `pip` installations afterward.
- It is helpful to mention file locks. Another MacOS text editor to mention is probably Atom, which I think is more popular than Brackets, and also doesn't lock files.

On MacOS, I came across several issues with the `opencv` image viewer windows that are used as a part of the `phenopype` workflow:

1. When opening an `opencv` image viewer window in general, there is an error and warning:

    > Class RunLoopModeTracker is implemented in both /Users/Seth/anaconda3/envs/pp/lib/python3.7/site-packages/cv2/.dylibs/QtCore (0x11500e7f0) and /Users/Seth/anaconda3/envs/pp/lib/libQt5Core.5.9.7.dylib (0x1282d1a80). One of the two will be used. Which one is undefined.
    QObject::moveToThread: Current thread (0x7fdc9d636fd0) is not the object's thread (0x7fdca22351f0).
    Cannot move to target thread (0x7fdc9d636fd0)

    It continues:
    >You might be loading two sets of Qt binaries into the same process. Check that all plugins are compiled against the right Qt binaries. Export DYLD_PRINT_LIBRARIES=1 and check that only one set of binaries are being loaded.
    QObject::moveToThread: Current thread (0x7fdc9d636fd0) is not the object's thread (0x7fdca22351f0).
    Cannot move to target thread (0x7fdc9d636fd0)

2. Then, once Issue (1) was solved: When using a `jupyter` notebook, when I attempt to view an image in an `opencv` window, the window opens, but it accepts no mouse input to zoom or scroll. As soon as I press any key, I get a spinning ball and a frozen window, and then I have to force quit and the kernel crashes.

3. Also, once Issue (1) was solved: In typical usage, `phenopype` image viewer windows are supposed to allow the user to zoom and scroll. The zooming and scrolling functionality is not present for me, regardless of how I run the code. (Actually, at the very beginning of my code review I was able to briefly open zoomable/scrollable windows, but now I cannot replicate that condition, despite trying a fairly wide range of approaches)

In the end, I was able to solve issue (1), more or less, but the other issues remained. It was possible for me to run the tutorial and example code from the command line or from standalone scripts. This was fine for the purposes of doing the review. The rest of my review continues below after this section. In case it would be of use for someone better-equipped than myself to solve this issue, here some things I learned about these issues:

- Issue (1) arise in the otherwise clean Python environment where I have installed `phenopype`, but it does not occur in a clean environment where I only install `opencv`.
- Issue (1) is solved if I only install `phenopype` version 1.0.5 with `pip install phenopype==1.0.5` (skipping updates and hotfixes).
- Issue (1) occurs if I do `pip install phenopype` and I also install the updates and hotfixes.
- Issue (1) occurs if I do `pip install phenopype` and I skip the updates and hotfixes.
- Issue (1) occurred for me in MacOS Big Sur and Catalina.
- Issue (1) crops up for other folks in other contexts. See [here](https://stackoverflow.com/questions/60271645/qt-complaining-2-sets-of-binaries), for instance. The advice in that SE thread is to install a version of `opencv` without any GUI functionality: `pip install opencv-python-headless`. If I do so, even without uninstalling `opencv` first, issue (1) is fixed. This was surprising. Why would installing an additional version of `opencv` without the capability of making image viewer windows solve the issue? Someone who understands dependencies better than me would probably know.
- Issues (2) and (3) can be replicated in the environment where `phenopype` was installed and also in an environment where only `opencv` was installed, by running a minimal example that uses `opencv` alone to open an image viewer window.
- Issue (2) occurs whether I attempt to run the notebook in a browser or within Atom (with a `jupyter` plugin installed).
- Issue (3) reduces the available precision of user input via mouse clicks on an image window (because it is not possible to zoom in), but I was still able to use the `phenopype` functions that take mouse clicks. Similarly, I was able to interact with the viewer via keystrokes (e.g. ESC to close a window, etc).

##### Overall documentation and user experience
In general, I found `phenopype` to be very well-documented, and I was impressed to see author has put together such a detailed set of tutorials and examples to illustrate how the package can be used. The package could stand by itself without so many detailed walkthroughs, however it is a big service to the community that there are walthroughs pitched at novices. I recommend "publication" of the package regardless of how the tutorials are updated, but if the author wants to polish the tutorials further, I have suggestions about their content and clarity. In the interest of keeping this post from becoming even more unwieldy in length, I will post those suggestions separately later today.
