### 2021-02-05 #24 @sdonoughe
### Notes and suggestions for `phenopype` tutorials

For the rest of my review, I've worked through the tutorials. Here are a set of suggestions, from minor typos to ideas about how to update the organization.

#### Tutorial 1
- "skript" -> "script"
- I found the usage of "bindings" here to be a little confusing.
- It is very thoughtful that this tutorial is here to introduce Python, and that there are also pointers to other introductions. My sense is that this intro might feel somewhat bewildering to a beginner. That is potentially okay, of course, since new things are almost always confusing. But I suspect the best way to make this tutorial maximally accessible would be to recruit an actual beginner and have them spend an hour or so trying to work through this section and then have them tell you how comfortable they feel about moving on to the next one. Since I'm not a beginner, I can't precisely point out where the hurdles are.
- There are several references to the "Resources" page. This is great! But there are a so many links there and it is not clear where someone should start. I suggest that, in general, the author take a slightly more authoritative tone in introducing material. "If you are new to Python, do [THIS] beginner overview first, then come back." Then next: "If have not done much image processing with code, read through [THIS] walkthrough second." And finally: "Okay now you are ready to begin with the `phenopype` tutorials."
- `os` is imported in cell 1 and cell 21. I don't know if that is intentional.
- Also, I realize all my references to cell numbers are might change. I think it will be good to restart the kernel and run through all cells in each notebook before adding them as browsable HTML to the `phenopype` website.
- "It is not informative neither possible" ->  "It is neither informative nor possible"
- I was somewhat puzzled as to the usefulness of opening lots of images at once. That is, since I couldn't think of why a user would want to do that, it might be better to demo a different functionlity as a way to illustrate the principles that are being taught for that section.
- I expected a link at the bottom of Tutorial 1 that takes me to Tutorial 2, but ultimately found one in the top nav bar.

#### Tutorial 2
- This is super minor: In the top table, capitalization is variable between rows and columns.
- "Fig. 2: Schematic of Phenopype’s prototyping workflow.": This figure only includes some of the functions that are in the code block immediately beneath it. I wondered: "what happened to the morphology operation?"
- Under Figure 2, in cell 10 there is an example sequence of image manipulation operations. As the user, I naturally ran the code, but there weren't any prompts in the tutorial or in the terminal telling me what to do. I think the reader needs something like:
> For the "draw mask" step, the user needs to select one or more regions of interest. Here we are using the "polygon" mode. Use left mouseclicks to trace an outline of some of the fish bones. Right click to undo the last segment. Then, when you're ready, press the "Control" key (or is better to write it as CTRL?) to end a given selection. Press ENTER key when you are done making region selections.

- Note to the author: I had to press CTRL twice to end a selection, but this does not appear to be the intended behavior from the docstring of the `create_mask` function. In general, in my hands, the `opencv` keystroke codes occasionally vary across operating systems.
- It is a bit confusing that the cell numbering restarts from 1 partway through the tutorial.
- For the Low-Throughput section, the output on the code block for me didn't match what was shown in the tutorial was:
```
- create mask
- include mask "mask1" pixels
- raw image
- contours saved under ../_temp/output/contours.csv (overwritten).
```
- The `contours.csv` file was present in the `_temp` directory, but it looks like it isn't being found?
- In the High-Throughput section, tell the user explicitly to press 'y' to make a new YAML file (I can see it in the terminal input/output lines, but I'm just trying to make it more approachable to a newcomer).
- The section on the `pype` is nice and detailed, but it feels like a lot to append to the end of this tutorial. I think the reader would be better served by splitting this tutorial into two, perhaps? It looks like this `pype` section sort of serves as an overview/preview of some of the material that also gets covered in the subsequent tutorial. That is fine! It just seemed like sort of a lot to digest in one go.
- This behavior did not occur for me: "Editing and saving the opened configuration file will close the image window, run the functions in the control file, and show the updated results." I think this is possibly due to the ongoing misbehavior of the `opencv` image viewer window.
- Also I could not test this functionality: "Closing the image window manually (with the X button in the upper right), also runs the functions in the control file, and show the updated results." ...because the "X" button is greyed out for me, presumably as a part of the same issue.
- Without these two things working, unfortunately I am missing out on a major part of the `phenopype` functionality---the ability to after the YAML file and observe live updates to the image output. If this affects other MacOS users too, it's a big enough deal that it probably warrants warning them ahead of time. In any case, it would be good if someone else on a Mac could do a very basic test of the image viewer functionality in order to better determine whether my own difficulties are in some way idiosyncratic to my machine itself.

#### Tutorial 3
- I imagine most readers would run the notebook from within the tutorials folder, which sets that as the working directory. It might confuse the reader that the working directory gets changed, which then breaks the relative path to the `images` dir in the `tutorials` directory. Is it necessary for the `_temp` folder to be where it is? Could it be in the tutorials folder itself, so that the reader doesn't have to update the path, and you could just go with the same relative paths for everyone?
- The `add_files` functionality is pretty slick!
- The block of text that begins "The remaining settings..." is pretty dense. I recommend using bullets or a table to organize it.
- All the embedded images that illustrate how `phenopype` works are much appreciated. It is nice touch. The black windows in "Step 2", however, are hard to read. That is, it's not clear what those represent.
- This cell runs indefinitely, even after I close the text file that gets opened automatically:
```
myproj.add_config(name = "lm",
                  config_preset="landmarks_plain",
                  interactive=True,
                  overwrite=True
                 )
```
- In the section on loading an existing project, again I had to update the path given in the tutorial, because it seems like the tutorial expects me to have the tutorial notebook to be running in a different directory? Or at least expects the working directory to be a different one. Here I had to update to:
```
import phenopype as pp
myproj = pp.project.load("../_temp/my_project/phenopype/")
```
- I suppose this means that it was expected that the working directory would have changed, but didn't, either because I skipped cells, or there was a change to the tutorial notebook. In any case, some other user might encounter this, so I bring it up.
- "controll" -> "control"

#### Tutorial 4
- I recommend omitting the option for users to load the previous project or make a new one, since it is a bit confusing with the commented out new directory line. Better, I think, to just have a minimal code block that generates an example project for this notebook
- Wow there are even walkthrough GIFs, very convenient!
- Again, the tutorial expects that we are working in a different directory than the one that actually has the `ipynb` files, I think? In the cell that follows this line: "We will use the image stickleback_side.jpg from the image folder in tutorials"
- The `add_scale` tool is very nice. It didn't completely work for me, or at least, it was unclear from the terminal prompts exactly what I should be doing to interact with it correctly.
- Specifically: when I run the function, I can click the two points for setting a reference length. Then I can enter a number (by the way, I would update the prompt to include the requested measurement units). But then it is not clear what I am I supposed to do next. The tutorial says to draw a rectangle around the scale image, but when I try to click on the image again the terminal prints: "already two points selected", and I don't know what to do next. Then pressing a key brings up red text on the image, and the terminal prints "- add column length". It is not clear what this means. I couldn't tell, in any case from the tutorial or the `add_scale` docstring. Bottom line: I think it would help to hit this function with a coat of polish, so to speak, and put a little more explanation in the tutorial and in its docstring.

#### Tutorial 5
- I couldn't assess all of the functions listed here, but I found this tutorial to be very clear.
- Consider putting this tutorial earlier in the series. Say, right after tutorial 1? I would have found it helpful to get these GUI basics before coming across image processing steps that involve the GUI in the other tutorials.
- Also, my personal opinion is that the GUI---and the capability to do semi-automated analysis that includes human interaction at more or more steps---is probably going to be the part of `phenopype` that will be most exciting and potentially useful to the research community, since it is a lot harder to build one's own GUI than it is to record a sequence of image transformations. I could be wrong about this, but if I'm right, maybe consider that a bit when describing the package on its landing page.

#### Tutorial 6
- This is super cool. My impression is that there _are_ several other published tools out there for specifically tracking animals in arenas (used for fish, flies, ants, birds, migrating animals, etc). [This](https://doi.org/10.1016/j.tree.2014.05.004) review is seven years old, but it suggests that a lot options were available. Maybe all the alternatives are too hard to use, or not open, or whatever, but still I think it would be helpful to explain how this tracking tool fits into the existing ecosystem.

#### Final thoughts
On the whole, these tutorials are comprehensive and careful, with a clear focus on making the tool accessible to people who are relatively new to coding in Python. The images and GIFs to illustrate the GUI are especially helpful.
