### 2021-02-07 #24 @mluerig
@sdonoughe and @agporto, thank you so much for those extensive and careful reviews - this is extremely helpful and I am eager to implement your suggestions! Here are some general responses to the issues and a roadmap for how I am planning to address them - @jbencook, @lwasser and @jlpalomino would be great to get your opinion on those, too:

### GUI

As I feared, the opencv GUI seems to be an issue for mac users, but can also potentially be unstable for windows/unix users. Since phenopype is built around opencv and its GUI, I probably can't do much more at this point other than trying to reduce internal sources of instability, better documentation of some of the potentially surprising behavior, and make it clear that phenopype currently doesn't stably run on macOS. Especially the latter is unfortunate, but currently the only solution I can think of, given that phenopype runs mostly stable on windows/unix.

### Tutorials

I looks like the tutorials need a bit of reorganizing, and both of you have given a lot of specific suggestions on how to do that. In addition to those specific changes, do you think it would make sense to move the tutorials (not the documentation) do a separate repo (e.g. "phenopype_tutorials") that is more self-contained and that people can just fork by itself? I am not sure what the best practices are, but as I was writing the tutorials and examples I felt that I am cluttering up the main program repo, so maybe a separation is useful

### Readme

I will add some citation information, and a comparison to other packages that @sdonoughe pointed at in his review. A statement that, at its core, phenopype is mostly a wrapper for other, more low-level CV libraries might also be helpful to users. Additionally, if people feel that this is needed, it I could add a little demonstration in the readme before pointing to external documentation and the tutorials.

So, I will work on this over the next few weeks and try to have a new version ready by the end of march or sometime in April.

@agporto and @sdonoughe, would you be available for a second round of reviews?

@jbencook, @lwasser and @jlpalomino, does all of this make sense or have I forgotten something?
