### 2021-02-08 #24 @sdonoughe
Yes, I am available for a second round of review, and without the big time delay.
