### 2021-02-09 #24 @jbencook
I think that plan makes sense @mluerig. Re: MacOS, I think it's very reasonable to not support GUI functionality as long as it's clear in the documentation.
