+++ testing/proj_test_new1/pype_config_MOD_v1.yaml
@@ -1,0 +1,7 @@
+info:
+  config_name: v1
+  template_name: test1.yaml
+  template_path: E:\git_repos\phenopype\phenopype\templates\test1.yaml
+  date_created: '20210216132548'
+  date_last_modified:
+steps: E:\git_repos\phenopype\phenopype\templates\test1.yaml
