+++ testing/proj_test_new1/data/0__stickleback_top/attributes.yaml
@@ -1,0 +1,14 @@
+image_original:
+  filename: stickleback_top.jpg
+  width: 3000
+  height: 2000
+  filepath: E:\Python\phenopype_dev\images\stickleback_top.jpg
+  filetype: .jpg
+image_phenopype:
+  date_added: '20210216132541'
+  mode: replica
+  resize: false
+  resize_factor: 1
+  filename: replica_stickleback_top.tif
+  width: 3000
+  height: 2000
