+++ testing/proj_test_new1/attributes.yaml
@@ -1,10 +1,0 @@
-project_info:
-  date_created: '20210216132530'
-  date_changed: '20210216132530'
-  phenopype_version: 2.0.dev0
-project_data:
-- 0__stickle1
-- 0__stickle2
-- 0__stickle3
-- 0__stickleback_side
-- 0__stickleback_top
