+++ testing/proj_test_new1/data/0__stickleback_side/attributes.yaml
@@ -1,14 +1,0 @@
-image_original:
-  filename: stickleback_side.jpg
-  width: 3000
-  height: 2000
-  filepath: E:\Python\phenopype_dev\images\stickleback_side.jpg
-  filetype: .jpg
-image_phenopype:
-  date_added: '20210216132540'
-  mode: replica
-  resize: false
-  resize_factor: 1
-  filename: replica_stickleback_side.tif
-  width: 3000
-  height: 2000
