+++ testing/proj_test_new1/data/0__stickle1/attributes.yaml
@@ -1,14 +1,0 @@
-image_original:
-  filename: stickle1.JPG
-  width: 2400
-  height: 1600
-  filepath: E:\Python\phenopype_dev\images\stickle1.JPG
-  filetype: .JPG
-image_phenopype:
-  date_added: '20210216132540'
-  mode: replica
-  resize: false
-  resize_factor: 1
-  filename: replica_stickle1.tif
-  width: 2400
-  height: 1600
