+++ source/readme.rst
@@ -1,75 +1,102 @@
-.. role:: raw-html-m2r(raw)
-   :format: html
+|image1|
 
+--------------
 
-Phenopype: a phenotyping pipeline for Python
+phenopype: a phenotyping pipeline for Python
 ============================================
 
-**Author:** `Moritz Lürig <https://luerig.net>`_\ :raw-html-m2r:`<br>`
-**License:** `LGPL <https://opensource.org/licenses/LGPL-3.0>`_  
++-----------+-----------+-----------+-----------+-----------+-----------+
+| Project   | Windows   | Linux     | OSX build | Coverage  | Style     |
+| status    | build     | build     |           |           |           |
++===========+===========+===========+===========+===========+===========+
+| |Project  | |Build    | |Build    | *none*    | |Coverage | |Code     |
+| Status:   | status|   | Status|   |           | Status|   | style|    |
+| Active.|  |           |           |           |           |           |
++-----------+-----------+-----------+-----------+-----------+-----------+
 
-.. list-table::
-   :header-rows: 1
+| **Author:** `Moritz Lürig <https://luerig.net>`__
+| **License:** `LGPL <https://opensource.org/licenses/LGPL-3.0>`__
 
-   * - Project status
-     - Windows build
-     - Linux build
-     - OSX build
-     - Coverage
-     - Style
-   * - 
-     .. image:: http://www.repostatus.org/badges/latest/active.svg
-        :target: http://www.repostatus.org/#active
-        :alt: Project Status: Active – The project has reached a stable, usable state and is being actively developed.
-     
-     - 
-     .. image:: https://ci.appveyor.com/api/projects/status/4o27rpjbe8ij2kj3?svg=true
-        :target: https://ci.appveyor.com/project/mluerig/phenopype
-        :alt: Build status
-     
-     - 
-     .. image:: https://travis-ci.org/mluerig/phenopype.svg?branch=master
-        :target: https://travis-ci.org/mluerig/phenopype
-        :alt: Build Status
-     
-     - *none*
-     - 
-     .. image:: https://coveralls.io/repos/github/mluerig/phenopype/badge.svg?branch=master
-        :target: https://coveralls.io/github/mluerig/phenopype?branch=master
-        :alt: Coverage Status
-     
-     - 
-     .. image:: https://img.shields.io/badge/code%20style-black-000000.svg
-        :target: https://github.com/psf/black
-        :alt: Code style
-     
+What is phenopype?
+------------------
 
+phenopype is a high throughput phenotyping pipeline for Python to
+support ecologists and evolutionary biologists in extracting high
+dimensional phenotypic data from digital images. phenopype’s workflow
+combines computer vision functions (using
+`opencv-python <https://github.com/opencv/opencv-python>`__ as a
+backbone) and project management tools to facilitate rapid data
+collection.
 
+|image2|
 
-.. image:: https://raw.githubusercontent.com/mluerig/phenopype/master/source/phenopype_logo.png
-   :target: https://raw.githubusercontent.com/mluerig/phenopype/master/source/phenopype_logo.png
-   :alt: 
+Features
+^^^^^^^^
 
+-  preprocessing functions () , segmentation and data extraction
+-  basic project management routines that can automatically organize
+   image data and create customizable analysis-templates.
+-  After finishing a project, users can share or archive the project
+   structure so that anyone can reproduce all collected data with only a
+   few lines of code (suitable for repositories like DRYAD or OSF).
 
-Package description
--------------------
+Why phenopype
+^^^^^^^^^^^^^
 
[…]
