### 2021-03-11 mluerig/phenopype#9 @mluerig
macOS: various window issues make phenopype unuseable
there are various issues with OpenCV's HighGUI module that are unrelated to phenopype. while the windows will typically open, they may not close at all, effectively crashing the Python kernel. here is an extensive description from of the code reviewers [1]:

> On MacOS, I came across several issues with the opencv image viewer windows that are used as a part of the phenopype workflow:
>
> When opening an opencv image viewer window in general, there is an error and warning:
>
> > Class RunLoopModeTracker is implemented in both /Users/Seth/anaconda3/envs/pp/lib/python3.7/site-packages/cv2/.dylibs/QtCore (0x11500e7f0) and /Users/Seth/anaconda3/envs/pp/lib/libQt5Core.5.9.7.dylib (0x1282d1a80). One of the two will be used. Which one is undefined.
> QObject::moveToThread: Current thread (0x7fdc9d636fd0) is not the object's thread (0x7fdca22351f0).
> Cannot move to target thread (0x7fdc9d636fd0)
>
> It continues:

> > You might be loading two sets of Qt binaries into the same process. Check that all plugins are compiled against the right Qt binaries. Export DYLD_PRINT_LIBRARIES=1 and check that only one set of binaries are being loaded.
> QObject::moveToThread: Current thread (0x7fdc9d636fd0) is not the object's thread (0x7fdca22351f0).
> Cannot move to target thread (0x7fdc9d636fd0)
>
> Then, once Issue (1) was solved: When using a jupyter notebook, when I attempt to view an image in an opencv window, the window opens, but it accepts no mouse input to zoom or scroll. As soon as I press any key, I get a spinning ball and a frozen window, and then I have to force quit and the kernel crashes.
>
> Also, once Issue (1) was solved: In typical usage, phenopype image viewer windows are supposed to allow the user to zoom and scroll. The zooming and scrolling functionality is not present for me, regardless of how I run the code. (Actually, at the very beginning of my code review I was able to briefly open zoomable/scrollable windows, but now I cannot replicate that condition, despite trying a fairly wide range of approaches)
>
> In the end, I was able to solve issue (1), more or less, but the other issues remained. It was possible for me to run the tutorial and example code from the command line or from standalone scripts. This was fine for the purposes of doing the review. The rest of my review continues below after this section. In case it would be of use for someone better-equipped than myself to solve this issue, here some things I learned about these issues:
>
> Issue (1) arise in the otherwise clean Python environment where I have installed phenopype, but it does not occur in a clean environment where I only install opencv.
> Issue (1) is solved if I only install phenopype version 1.0.5 with pip install phenopype==1.0.5 (skipping updates and hotfixes).
> Issue (1) occurs if I do pip install phenopype and I also install the updates and hotfixes.
> Issue (1) occurs if I do pip install phenopype and I skip the updates and hotfixes.
> Issue (1) occurred for me in MacOS Big Sur and Catalina.
> Issue (1) crops up for other folks in other contexts. See here, for instance. The advice in that SE thread is to install a version of opencv without any GUI functionality: pip install opencv-python-headless. If I do so, even without uninstalling opencv first, issue (1) is fixed. This was surprising. Why would installing an additional version of opencv without the capability of making image viewer windows solve the issue? Someone who understands dependencies better than me would probably know.
> Issues (2) and (3) can be replicated in the environment where phenopype was installed and also in an environment where only opencv was installed, by running a minimal example that uses opencv alone to open an image viewer window.
> Issue (2) occurs whether I attempt to run the notebook in a browser or within Atom (with a jupyter plugin installed).
> Issue (3) reduces the available precision of user input via mouse clicks on an image window (because it is not possible to zoom in), but I was still able to use the phenopype functions that take mouse clicks. Similarly, I was able to interact with the viewer via keystrokes (e.g. ESC to close a window, etc).

[1] https://github.com/pyOpenSci/software-review/issues/24#issuecomment-773319194
