+++ source/api.rst
@@ -5,9 +5,7 @@ API reference
 Projects
 --------
 
-Phenopype projects are composed of a directory tree in which each folder contains the copy or a link to a single raw image file. Each raw 
-should have only one folder, so that result or intermediate output from different iterations of processing and analysis are stored side 
-by side. The project root folder should be separate from the raw data, e.g. as a folder inside of your project folder:
+Phenopype projects are composed of a directory tree in which each folder contains the copy or a link to a single raw image file. Each raw should have only one folder, so that result or intermediate output from different iterations of processing and analysis are stored side by side. The project root folder should be separate from the raw data, e.g. as a folder inside of your project folder:
 
 .. code-block:: python
 
@@ -26,7 +24,7 @@ Above code creates a folder structure as follows:
 .. code-block:: bash
 
 	my_project\
-	 data_raw 
+	 data_raw
 	 phenopype\		# create phenopype project here with "pp.project"
 	  data\ 		# created automatically
 	   project.data		# created with "pp.project.save"
@@ -37,7 +35,7 @@ Above code creates a folder structure as follows:
 	    results_v1.csv
 	   file2\
 	    ...
-	   ... 
+	   ...
 	 data
 	 scripts
 	 manuscript
@@ -45,9 +43,7 @@ Above code creates a folder structure as follows:
 
 .. important::
 
-	Currently the only useful information contained in the project object (:code:`myproj`) is a list of all directories inside the
-	project's directory tree. It is important to save `both` the project AND all results in the data directories using the appropriate 
-	functions in (:ref:`Export`). Future release will expand the functionality of the project class and its associated methods. 
+	Currently the only useful information contained in the project object (:code:`myproj`) is a list of all directories inside the project's directory tree. It is important to save `both` the project AND all results in the data directories using the appropriate functions in (:ref:`Export`). Future release will expand the functionality of the project class and its associated methods.
 
 .. autoclass:: phenopype.main.project
 	:members:
@@ -61,11 +57,132 @@ pype (high throughput function)
 
 .. raw:: html
 
-	Method used to process image datasets with high throughput and reproducibility. Should be used in 
-	conjunction with image organized in a Phenopype directory structure. To get started with the high 
-	througput workflow, see <a href="tutorial_2_phenopype_workflow.html">Tutorial 2</a> and 
-	<a href="tutorial_3_managing_projects_1.html">Tutorial 3</a>.
+	Method used to process image datasets with high throughput and reproducibility. Should be used in conjunction with image organized in a Phenopype directory structure. To get started with the high througput workflow, see <a href="tutorial_2_phenopype_workflow.html">Tutorial 2</a> and <a href="tutorial_3_managing_projects_1.html">Tutorial 3</a>.
 	<br>
+
+
+.. important::
+
+		``pype`` behavior
+	~~~~~~~~~~~~~~~~~
+
+	The ``pype`` function has specific implicit behavior that aims at
+	supporting speed and robustness when working in "production" (i.e. when
+	performing the actual analysis of large image datasets compared to
+	prototyping and low throughput workflow). Here I list some important
+	aspects of that behavior.
+
+	Window control
+	^^^^^^^^^^^^^^
+
+	The ``pype`` method uses HighGUI windows, which can be sometimes
+	unstable - be sure to check `Tutorial
+	2 <tutorial_2_phenopype_images.ipynb#Window-control>`__. If your text
+	window didn't show up when using the ``pype`` function, make sure you
+	selected a default text editor to handle YAML files (`see the
[…]
