+++ source/api.rst
@@ -57,131 +57,47 @@ pype (high throughput function)
 
 .. raw:: html
 
-	Method used to process image datasets with high throughput and reproducibility. Should be used in conjunction with image organized in a Phenopype directory structure. To get started with the high througput workflow, see <a href="tutorial_2_phenopype_workflow.html">Tutorial 2</a> and <a href="tutorial_3_managing_projects_1.html">Tutorial 3</a>.
-	<br>
-
+	Method used to process image datasets with high throughput and reproducibility. Should be used in conjunction with image organized in a Phenopype directory structure. 	The ``pype`` function has specific implicit behaviour that aims at supporting speed and robustness when working in "production" (i.e. when performing the actual analysis of large image datasets compared to	prototyping and low throughput workflow). Here I list some important aspects of that behaviour.
 
 .. important::
 
-		``pype`` behavior
-	~~~~~~~~~~~~~~~~~
+	**Enhanced Window control**
 
-	The ``pype`` function has specific implicit behavior that aims at
-	supporting speed and robustness when working in "production" (i.e. when
-	performing the actual analysis of large image datasets compared to
-	prototyping and low throughput workflow). Here I list some important
-	aspects of that behavior.
+	In addition to regular window control functions documented in `(Tutorial 2) <https://mluerig.github.io/phenopype/tutorial_2_phenopype_images.html#Window-control>`__:
 
-	Window control
-	^^^^^^^^^^^^^^
+	-  Editing and saving the opened configuration file in the text editor will trigger another iteration, i.e. close the image window and run the config file again.
+	-  Closing the image window manually (with the X button in the upper right), also runs triggers another run.
+	-  ``Esc`` will close all windows and interrupt the pype routine (triggers ``sys.exit()``, which will also end a Python session if run from the command line), as well as any loops.
+	-  Each step that requires user interaction (e.g. ``create_mask`` or ``landmarks``) needs to be confirmed with ``Enter`` until the next function in the sequence is executed.
+	-  At the end of the analysis, when the final steps (visualization and export functions) have run, use  ``Ctrl+Enter`` to finish and close the window.
 
-	The ``pype`` method uses HighGUI windows, which can be sometimes
-	unstable - be sure to check `Tutorial
-	2 <tutorial_2_phenopype_images.ipynb#Window-control>`__. If your text
-	window didn't show up when using the ``pype`` function, make sure you
-	selected a default text editor to handle YAML files (`see the
-	Installation Instructions <installation.html#choose-a-text-editor>`__).
-	Here is another summary of the window behavior when working with
-	``pype`` in high throughput workflow:
-
-	-  Editing and saving the opened configuration file in the text editor
-	   will trigger another ``pype`` iteration, i.e. close the image window,
-	   run the functions in the control file, and display the updated
-	   results.
-	-  Closing the image window manually (with the X button in the upper
-	   right), also runs the functions in the control file, and show the
-	   updated results.
-	-  ``Esc`` will close all windows and interrupt the pype routine
-	   (``Esc`` triggers ``sys.exit()``, which will also end a Python
-	   session if run from the command line) [1,2] , as well as any loops
-	   (`see Tutorial
-	   4 <tutorial_4_managing_projects.ipynb#Using-pype-with-project-folders>`__).
-	-  Each step that requires user interaction (e.g. ``create_mask`` or
-	   ``landmarks``) needs to be confirmed with ``Return`` until the next
-	   function in the sequence is executed [1,2].
-	-  At the end of the analysis, when the final steps (visualization and
-	   export functions) have run, you can end the pype routine with another
[…]
