### 2021-03-14 #24 @mluerig
Hi again!

So, below are the responses to your comments, I have updated phenopype accordingly, it is now version 2.0.0. Since v1.0.5, which was the one I originally submitted, I have changed a few things that how users the yaml configuration files (I moved to list structures for greater flexibility when configuring the high throughput workflow). Therefore, it is not backwards compatible and thus got a major version update. I also added some minor code changes in the background that improve code readability, but should not affect user experience. All changes are documented in the changelog.

I modified what I could, but some things (failing on macOS, HighGUI instability, changing how annotations are stored) were beyond the scope of these revisions (as both of you already acknowledged in your reviews). However I'd be very happy to have a chat with you about what can be done in the future.

Below I reply to your comments - hope they make sense, otherwise let me know.

Arthur
======

> The documentation is quite extensive and well organized, but I think one thing that would help new users is moving the Tutorial on GUI interactions (currently, Tutorial 5) to the beginning. This tutorial explains in more detail how to interact with OpenCV's High GUI module. In my experience with the current package, this module was by far the hardest one to interact with. It is somewhat unstable, crashing the jupyter kernel with some frequency, and its keystrokes are not intuitive. Given that this module is important in mediating user interactions with Phenopype, I think it would be useful to have a tutorial on it front and center. It might prevent new users from getting overly frusfrated and ending up abandoning the package.

I followed your (and Seth's suggestion) and rearranged the tutorials: tutorial 2 now covers GUI interactions. I completely overhauled this tutorial and I now also discuss at the most common HighGUI issues in OpenCV, so that users are prepared and can try the workarounds I suggested. Indeed, in the long term I will have to consider alternatives to the OpenCV GUIs.

> Another aspect that might need some attention is the standardization of the root directory. When going through the tutorials, perhaps it would be useful to explicitely set the root directory, since the root directory jumps around a little across the different tutorials (sometimes it starts on one place, sometimes on another)

This should be fixed: every tutorial now assumes the folder `tutorials` in the downloaded `phenopype-master` folder as root.

> When running Tutorial 2 and going through the 'Low throughput workflow' I get the following error:

fixed: https://github.com/mluerig/phenopype/issues/7

> Similarly, when running Tutorial 5 and going through the polylines example, I get the following error:

fixed: https://github.com/mluerig/phenopype/issues/8

> Having followed the success of the COCO (https://cocodataset.org/#home) data format, I wonder if phenopype would benefit from having a similar architecture for data storage. By storing different types of annotations in a single JSON file, COCO has created an almost universal file format that makes it remarkably easy to train different machine learning algorithms using the same dataset (whether trying to predict landmarks or masks, for example). Again, I do not expect this change to occur, but I think it might be worth considering in the future.

This is something I would love to chat about. I had a look at the COCO json files, and I think I would like to follow a similar, albeit not identical structure (I think phenotypic data would need to have somewhat different structure than pure segmentation labels).

Seth
====

> Brief demonstration usage: There currently isn't a demonstration usage directly in the README, but there are, of course, lots of links to demonstrations. If a demo is needed here, perhaps the best would be have code that opens an image, takes user mouse clicks to select a region, and then does, say, a threshold operation on the contents of that region, then displays and saves the results.

I followed your suggestion and added a gif to the readme that shows some basic image processing operations and the results in real time.

>How the package compares to similar packagese: I think it would be useful to have an explanation for how this package relates to the most widely used Python image manipulation packages, and why a user might want to use phenopype in combination with one or more of those packages. As for the graphical interface for taking user input interactively, the closest Python tool that I'm aware of is napari, even though it is still in alpha. For users that do not have much coding experience, another option for them is to use Fiji/ImageJ, and then record a macro of their image processing steps. Here, phenopype offers functionality that is not present in any of these alternatives, so it would be useful to point that out too.

I used some of what you wrote below (actually, almost one-to-one, so you should get credit) in the review to point out that phenopype is a wrapper for OpenCV in combination with a project management ecosystem.

> Citation information: If/when this package is described in a journal article, then that citation can be added, but even in the meantime I can see biologists wanting to try phenopype out and potentially even using it in publications. So I'd definitely recommend having a "how to cite phenopype" section.

I added citation instructions

> One might wonder: why wouldn't a biologist just use opencv by itself? Or scikit-image, or any other Python library of image processing tools, for that matter? Answer: Because phenopype is aiming to augment, rather than replace, the utility of existing CV libraries for scientists measuring phenotypes. Put differently, phenopype is not a library of granular image processing functions. Instead, it is a set of wrappers around some of opencv's most broadly used functions, combined with a simplified GUI for taking user input; these collectively act as a scientist's bookkeeper during an image processing project. In doing so, it fills a major void in the scientific Python toolkit. As far as I know there is no other Python tool that enables the user to create a pipeline of image-processing steps that incorporates manual input along the way. phenopype is also (to my knowledge) unique in defining a terse, yet legible format for recording and replicating the parameters of such a pipeline.

I actually think that this text gives a nice summary of what phenopype is providing, and how it is distinct from other packages. I took the liberty of using some of your formulations, and incorporating them into the readme.

> There are some tradeoffs involved in structuring the package this way. For one, the set of image-manipulation functions is impressive, but it understandably covers only a small subset of the image-parsing tasks that are regularly encountered in image analysis work for biological data. This is not a knock against phenotype in the slightest. But a user that wanted to integrate additional image-processing functions into the middle of a workflow would need to invest some time to work out how phenopype works under the hood if they wanted to take advantage of the replicability tools that are the core of the package.

So, I wonder if, in the long term, there might be a better way of wrapping OpenCV (and potentially scikit-image) functions, maybe even through means of a toolchain that does this in an automatic fashion. Would be great to hear your and @agporto's thoughts on this.

> Notes on the Installation page itself:

I decluttered and streamlined the installation instructions:
- I made it clear that managing environments via `conda` and installing phenopype via `pip` are the default. In addition, I provided some more information regarding Python installation and package management, as this appears to be a common source of confusion.
- using `Spyder` is, of course, optional - I now state this in a tip-bubble. I do think that many potential users will be coming from R/Rstudio, and are accustomed to that user-experience and workflow, so I want to make this a strong recommendation.
- I removed the options to install hotfixes and development versions, which carried over from my inexperience with the git workflow.
- I streamlined the information regarding text editors.

> On MacOS, I came across several issues with the opencv image viewer windows that are used as a part of the phenopype workflow:

I am currently not able to solve this, but in the long term I am planning to make phenopype usable on all platforms (including macOS) by either fixing the OpenCV GUI issues OR by finding other means of visualization and user interactions (napari for instance looks very promising).

Until then, to make it clear that phenopype doesn't run on macOS, I added a [little section to the readme](https://github.com/mluerig/phenopype#important-information) and documented your efforts by transferring them to a phenopype [issue](https://github.com/mluerig/phenopype/issues/9)

---

### Tutorials

Note that tutorial 5 got moved to position 2 - I commenting here in the new order

#### Tutorial 1

> It is very thoughtful that this tutorial is here to introduce Python, and that there are also pointers to other introductions. My sense is that this intro might feel somewhat bewildering to a beginner. That is potentially okay, of course, since new things are almost always confusing. But I suspect the best way to make this tutorial maximally accessible would be to recruit an actual beginner and have them spend an hour or so trying to work through this section and then have them tell you how comfortable they feel about moving on to the next one. Since I'm not a beginner, I can't precisely point out where the hurdles are.

I added a little balloon to ask for feedback.

> There are several references to the "Resources" page. This is great! But there are a so many links there and it is not clear where someone should start. I suggest that, in general, the author take a slightly more authoritative tone in introducing material. "If you are new to Python, do [THIS] beginner overview first, then come back." Then next: "If have not done much image processing with code, read through [THIS] walkthrough second." And finally: "Okay now you are ready to begin with the phenopype tutorials."

Fully agree - I added a **Further Reading** balloon to this and all other tutorials (and will do so for the examples).

> I was somewhat puzzled as to the usefulness of opening lots of images at once. That is, since I couldn't think of why a user would want to do that, it might be better to demo a different functionlity as a way to illustrate the principles that are being taught for that section.

I sometimes like to look at several images to look for differences in traits, or to compare different thresholding outcomes - I thought it was useful, but if you really think it's confusing I can remove it.

> I expected a link at the bottom of Tutorial 1 that takes me to Tutorial 2, but ultimately found one in the top nav bar.

fixed - now all tutorials have links to the next one

#### Tutorial 2

> Consider putting this tutorial earlier in the series. Say, right after tutorial 1? I would have found it helpful to get these GUI basics before coming across image processing steps that involve the GUI in the other tutorials.

done - Tutorial 5 is now the second one

> Also, my personal opinion is that the GUI---and the capability to do semi-automated analysis that includes human interaction at more or more steps---is probably going to be the part of phenopype that will be most exciting and potentially useful to the research community, since it is a lot harder to build one's own GUI than it is to record a sequence of image transformations. I could be wrong about this, but if I'm right, maybe consider that a bit when describing the package on its landing page.

good point - I mentioned it now in the reworked readme

#### Tutorial 3

I cleaned this tutorial up a bit, i.e. I cut back on intro-text, removed the "morphology" operations, I added instructions for what the expected interactions are, and streamlined the end about yaml and pype (pype-behavor section got moved to API)

> "Fig. 2: Schematic of Phenopype’s prototyping workflow.": This figure only includes some of the functions that are in the code block immediately beneath it. I wondered: "what happened to the morphology operation?"

I removed morphology operation to keep it simple

> Under Figure 2, in cell 10 there is an example sequence of image manipulation operations. As the user, I naturally ran the code, but there weren't any prompts in the tutorial or in the terminal telling me what to do. I think the reader needs something like:

I added some instructions at the very beginning to show users what they have to do

> Note to the author: I had to press CTRL twice to end a selection, but this does not appear to be the intended behavior from the docstring of the create_mask function. In general, in my hands, the opencv keystroke codes occasionally vary across operating systems.

Since for this example only one polygon-mask is needed, I changed the text to "finish and close with Enter" to keep complexity low. Multi-polygon masks are also covered in Example 5

> For the Low-Throughput section, the output from one code block for me didn't match what was shown in the tutorial. This was the one:

should be fixed

> The contours.csv file was present in the _temp directory, but it looks like it isn't being found?

should be fixed

> In the High-Throughput section, tell the user explicitly to press 'y' to make a new YAML file (I can see it in the terminal input/output lines, but I'm just trying to make it more approachable to a newcomer).

this has changed

> The section on the pype is nice and detailed, but it feels like a lot to append to the end of this tutorial. I think the reader would be better served by splitting this tutorial into two, perhaps? It looks like this pype section sort of serves as an overview/preview of some of the material that also gets covered in the subsequent tutorial. That is fine! It just seemed like sort of a lot to digest in one go.

agreed - I sourced it out to the API.  I first thought to make another tutorial or section only for the pype, but I think it's better to keep all relevant information in one place, so the API made more sense to me. let me know if this helps or whether it complicates things.

> This behavior did not occur for me: "Editing and saving the opened configuration file will close the image window, run the functions in the control file, and show the updated results." I think this is possibly due to the ongoing misbehavior of the opencv image viewer window.

I think this is another macOS problem - sorry :-)

#### Tutorial 4

> I imagine most readers would run the notebook from within the tutorials folder, which sets that as the working directory. It might confuse the reader that the working directory gets changed, which then breaks the relative path to the images dir in the tutorials directory. Is it necessary for the _temp folder to be where it is? Could it be in the tutorials folder itself, so that the reader doesn't have to update the path, and you could just go with the same relative paths for everyone?

fixed - now all tutorials and examples start from phenopype/tutorials, and save output under phenopype/tutorials/_temp

> The block of text that begins "The remaining settings..." is pretty dense. I recommend using bullets or a table to organize it.

fixed

> All the embedded images that illustrate how phenopype works are much appreciated. It is nice touch. The black windows in "Step 2", however, are hard to read. That is, it's not clear what those represent.

I added some text to show that they are the YAML config files

> This cell runs indefinitely, even after I close the text file that gets opened automatically:

Those functions got reworked ... hope it works now

> In the section on loading an existing project, again I had to update the path given in the tutorial, because it seems like the tutorial expects me to have the tutorial notebook to be running in a different directory? Or at least expects the working directory to be a different one. Here I had to update to:

should be fixed

#### Tutorial 5

> I recommend omitting the option for users to load the previous project or make a new one, since it is a bit confusing with the commented out new directory line. Better, I think, to just have a minimal code block that generates an example project for this notebook

agreed, and done!

> Again, the tutorial expects that we are working in a different directory than the one that actually has the ipynb files, I think? In the cell that follows this line: "We will use the image stickleback_side.jpg from the image folder in tutorials"

no, the images are just a subfolder - the working directory shouldn't move anymore

> Specifically: when I run the function, I can click the two points for setting a reference length. Then I can enter a number (by the way, I would update the prompt to include the requested measurement units). But then it is not clear what I am I supposed to do next. The tutorial says to draw a rectangle around the scale image, but when I try to click on the image again the terminal prints: "already two points selected", and I don't know what to do next. Then pressing a key brings up red text on the image, and the terminal prints "- add column length". It is not clear what this means. I couldn't tell, in any case from the tutorial or the add_scale docstring. Bottom line: I think it would help to hit this function with a coat of polish, so to speak, and put a little more explanation in the tutorial and in its docstring.

I think this may have been a bug I fixed in the meantime - should be working now

#### Tutorial 6

> This is super cool. My impression is that there are several other published tools out there for specifically tracking animals in arenas (used for fish, flies, ants, birds, migrating animals, etc). This review is seven years old, but it suggests that a lot options were available. Maybe all the alternatives are too hard to use, or not open, or whatever, but still I think it would be helpful to explain how this tracking tool fits into the existing ecosystem.

I agree that the tracking module at this point seems more of an appendix. My goal is that, at some point, all the core processing functions can be used on objects found in every single frame.
