### 2021-04-08 fb36543
fix draw silent mode
