### 2021-04-30 #24 @lwasser
@agporto  @sdonoughe can you please have a look at @mluerig responses to your review above? are you  satisfied with the responses? the one issue i see is the  MAC compatibility issue. i am going to see how ropensci handles these types of issues in a review and will respond accordingly. That aside  are you  both   content  with t he review r esponse  above OR are  there  additional items that you'd like to see addressed?
my apologies for  the huge  delay  - a lot has been going on here locally (not work related) and it's really thrown me off.
