### 2021-05-03 #24 @sdonoughe
### Three main notes

1. The reorganization and revision of the tutorials is very impressive. Even before my review they were already quite clear, and now they're even better.
2. As before, I couldn't get the image windows to behave as intended on MacOS, so I wasn't able to actually test all the functionality.
3. That said, I read through the tutorials, and I found them very thoughtfully structured and explained. All I found were a few miscellaneous typos:

#### Tutorial 1
I think this is a perfectly pitched intro.

#### Tutorial 2
Overall, I find the updated tutorials to be extremely thorough and helpful, and this one in particular.

A small typo: "Currently Phenopype uses the standard HighGUI libraries that ship with the most recent precompiled opencv-contrib-python package that is listed on the, which is Qt for Linux and macOS, and Win32 UI on Windows."

Something is missing after "listed on the"

#### Tutorial 3
Small typo: "To learn how to analyze **entir** data sets with the high-throughput method, move on to the next Tutorial 4."

#### Tutorial 4
Small typo: "initiatlized"

#### Tutorial 5
I found this very useful! I think it would be helpful to add a small summary bit at the end, like you've done for the previous tutorials.

#### Tutorial 6
Small typo: "exlcude"
Small typo: "applied matters The following"

### Bottom line

@lwasser I'm totally content with the review response.

Congratulations on a comprehensive and useful tool, @mluerig!
