+++ phenopype/utils_lowlevel.py
@@ -49,7 +49,6 @@ class _image_viewer:
         """
         ## kwargs
         self.flag_tool = kwargs.get("tool", None)
-        self.flag_draw = kwargs.get("draw", False)
         self.flag_zoom_mode = kwargs.get("zoom", "continuous")
         self.zoom_mag = kwargs.get("mag", 0.7)
         self.zoom_n_steps = kwargs.get("steps", 20)
@@ -138,9 +137,13 @@ class _image_viewer:
                 self.text_size = kwargs.get("label_size", _auto_text_size(image))
                 self.text_width = kwargs.get("label_size", _auto_text_width(image))
                 self.label_colour = colours[kwargs.get("label_colour", "black")]
-        if self.flag_draw:
-            self.line_colour = colours[kwargs.get("line_colour", "black")]
-
+            elif self.flag_tool == "draw":
+                self.drawing = False
+                self.point_size_orig = kwargs.get("point_size", _auto_point_size(image))
+                self.point_size = copy.deepcopy(self.point_size_orig)
+                self.point_colour = colours[kwargs.get("point_colour", "red")]
+                self.orig_contours = kwargs.get("contours")
+                self.zoom_flag = True
 
         ## this needs to be decluttered. there should only be one type of list 
         ## (points, not rectangle separately). relevant parameters for the 
@@ -197,32 +200,55 @@ class _image_viewer:
         cv2.startWindowThread()  ## needed for Mac OS ??
         cv2.setMouseCallback(self.window_name, self._on_mouse_plain)
         cv2.resizeWindow(self.window_name, self.canvas_width, self.canvas_height)
-        cv2.imshow(self.window_name, self.canvas)
-
+        
         ## window control
+        self.keypress = None
         if window_control == "internal":
-            if self.flag_test_mode == True:
-                self.done = True
-                self.finished = True
-                cv2.waitKey(self.wait_time)
-                cv2.destroyAllWindows()
-            elif self.flag_test_mode == False:
-                if cv2.waitKey() == 13:
-                    self.done = True
-                    cv2.destroyAllWindows()
-                    # for i in range(10):
-                    #     cv2.waitKey(1)
-                elif cv2.waitKey() == 10:
+            
+            while not any([self.finished, self.done]):
+                cv2.imshow(self.window_name, self.canvas)
+                self.keypress = cv2.waitKey(500)     
+                if self.flag_test_mode == True:
                     self.done = True
                     self.finished = True
+                    cv2.waitKey(self.wait_time)
                     cv2.destroyAllWindows()
-                    # for i in range(10):
-                    #     cv2.waitKey(1)
-                elif cv2.waitKey() == 27:
-                    cv2.destroyAllWindows()
-                    # for i in range(10):
-                    #     cv2.waitKey(1)
-                    sys.exit("\n\nTERMINATE (by user)")
+                elif self.flag_test_mode == False:
+                    if self.keypress == 13:
+                        self.done = True
+                        cv2.destroyAllWindows()
+                    elif self.keypress == 10:
+                        self.done = True
+                        self.finished = True
+                        cv2.destroyAllWindows()
+                    elif self.keypress == 27:
+                        cv2.destroyAllWindows()
+                        sys.exit("\n\nTERMINATE (by user)")
+                        print("end while")     
+                    ## remove last points
+                    elif self.keypress == 26:
[…]
