### 2021-05-10 7df81c1
fix examples v2/typos in tutorials
