+++ CHANGELOG.txt
@@ -9,6 +9,7 @@ bugfix:
 	- save_masks as df would take ndarray instead of DataFrame
 	- select_canvas with ndarray would not allow to select raw
 - draw_landmarks now takes label=False/True and offset argument
+- updated and checked all examples for errors
 
 2.0.0
 -----
