+++ _sources/index.rst.txt
@@ -1,0 +1,183 @@
+.. include:: readme.inc
+
+.. raw:: html
+
+	<hr>
+
+	<div class="container-fluid">
+	<div class="row">
+	<div class="col-md-6">
+	<h2>Documentation</h2>
+
+.. toctree::
+	:maxdepth: 2
+
+	installation
+	api
+	resources
+
+.. raw:: html
+
+	</div>
+	<div class="col-md-6">
+	<h2>Tutorials</h2>
+
+.. toctree::
+	:maxdepth: 1
+
+	tutorial_0
+	tutorial_1_python_intro
+	tutorial_2_phenopype_images
+	tutorial_3_phenopype_workflows
+	tutorial_4_managing_projects
+	tutorial_5_references
+	tutorial_6_video_analysis
+
+.. raw:: html
+
+	</div>
+	</div>
+	<hr>
+
+
+Examples
+--------
+
+.. raw:: html
+
+	<div class="row">
+
+		<div class="col-md-4">
+			<div class="thumbnail">
+			<a href="example_1_detect_objects_isopods.html">
+
+.. image:: thumbs/ex1.jpg
+
+.. raw:: html
+
+			<b><p style="text-decoration: underline;">Detecting aquatic isopods</p></b>
+			<b>Organism: </b>waterlouse (or sowbug)<br>
+			<b>Species: </b><i>Asellus aquaticus</i> <br>
+			<b>Traits: </b>size, pigmentation, body-shape <br>
+			</a>
+			</div>
+		</div>
+
+		<div class="col-md-4">
+			<div class="thumbnail">
+			<a href="example_2_landmarks_stickleback.html">
+
+.. image:: thumbs/ex2.jpg
+
+.. raw:: html
+
+			<b><p style="text-decoration: underline;">Placing landmarks</p></b>
+			<b>Organism: </b>threespine stickleback <br>
+			<b>Species: </b><i>Gasterosteus aculeatus</i> <br>
+			<b>Traits: </b>landmarks, geometric morphometrics <br>
+			</a>
+			</div>
+		</div>
+
+		<div class="col-md-4">
+			<div class="thumbnail">
+			<a href="example_3_phytoplankton.html">
+
+.. image:: thumbs/ex3.jpg
+
+.. raw:: html
+
+			<b><p style="text-decoration: underline;">Detecting plankton cells</p></b>
+			<b>Organism: </b>phytoplankton <br>
+			<b>Species: </b><i>community</i> <br>
+			<b>Traits: </b>pixel level intensity, shape-metrics <br>
+			</a>
+			</div>
+		</div>
+
+	</div>
+	<div class="row">
+
+		<div class="col-md-4">
+			<div class="thumbnail">
+			<a href="example_4_video_analysis_stickleback.html">
+
+.. image:: thumbs/ex4.jpg
+
+.. raw:: html
+
+			<b><p style="text-decoration: underline;">Recording predator prey interactions</p></b>
+			<b>Organism: </b>stickleback and isopods <br>
+			<b>Species: </b><i>G. aculeatus, A. aquaticus</i> <br>
+			<b>Traits: </b>movement, size <br>
+			</a>
+			</div>
+		</div>
+
+		<div class="col-md-4">
+			<div class="thumbnail">
+			<a href="example_5_shape_stickleback.html">
+
+.. image:: thumbs/ex5.jpg
+
+.. raw:: html
+
+			<b><p style="text-decoration: underline;">Measuring armour plate area</p></b>
+			<b>Organism: </b>threespine stickleback <br>
+			<b>Species: </b><i>G. aculeatus</i> <br>
+			<b>Traits: </b>counting armor plates; shape and size<br>
+			</a>
+			</div>
[…]
