+++ _sources/example_3_phytoplankton.ipynb.txt
@@ -1,0 +1,599 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Example 3: Fluorescence intensity and shape of phytoplankton cells\n",
+    "\n",
+    "In this example we are trying to i) detect phytoplankton cells in 1-channel images (black and white images), ii) measure their pixel-intensities (=fluorescence intensities) and their their shape, and iii) work towards building a classifier for taxonomic groups of phytoplankton. \n",
+    "\n",
+    "The first two objectives are a rather low hanging fruit, as they involve classic computer vision approaches that phenopype can readily supply. The latter goal is more longterm and involves experimenting with classic shape feature detection as provided by phenopype's `shape_features` function (e.g. [image moments](https://docs.opencv.org/3.4/d8/d23/classcv_1_1Moments.html)), and potentially also building up towards a deep learning pipeline usign the contours detected with phenopype. I will update this example as work progresses.  \n",
+    "\n",
+    "Images kindly provided by Irene Gallego and Anita Narwani.\n",
+    "\n",
+    "\n",
+    "<div class=\"row; text-align: left\">\n",
+    "<div class=\"col-md-6\">\n",
+    "<img src=\"_assets/ex3_before.jpg\">\n",
+    "\n",
+    "**Input** - Phytoplankton fluorescence. Each sample includes 1 brightfield gray scale image and 3 fluorescence images at different wavelengths.\n",
+    "</div>\n",
+    "<div class=\"col-md-6\">\n",
+    "<img src=\"_assets/ex3_after.jpg\">\n",
+    "\n",
+    "**Results** - Phenopype `thresholding` function detects the contours of phytoplankton cells (green) and holes within (red).</div>\n",
+    "</div>\n"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Approach"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "All four images show the same objects, but with different pixel intensities. It may be useful to combine information from all images into the segmentation process. This is requires customization of existing workflows in Phenopype, for which the low throughput workflow is best suited: here we have access to intermediate output, and can also integrate low level functions from the `OpenCV` library. Refer to [Tutorial 2](tutorial_2_phenopype_workflow.ipynb) for mor information about the different workflows.\n",
+    "\n",
+    "The appraoch is to load all images separately, do segmentation separately, find contours create four binary masks from all contour coordinates, combine the binary masks into a single image, find contours on the binary image.\n",
+    "\n",
+    "\n",
+    "[1] https://docs.opencv.org/master/d0/d86/tutorial_py_image_arithmetics.html\n",
+    "\n",
+    "<center>\n",
+    "<div style=\"width:800px; text-align: left\" >\n",
+    "<img src=\"_assets/ex3_phyto_layers.png\">\n",
+    "    \n",
+    "**Fig. 1:** Each sample includes four images: 1 brightfield (top) and 3 fluorescence measurements (black, bottom). Due to different pigments, not all spcecies are visible in each image, because of different emission spectra. For example, the two long string shaped cells only occur in two of the fluorescence channels, but not the third one or the brightfield. Therefore, including information from all images for segmentation may be useful to get close to the \"true\" cell count and community composition within a sample. \n",
+    "    \n",
+    "</div>\n",
+    "</center>"
+   ]
[…]
