+++ phenopype/_version.py
@@ -2,4 +2,4 @@
 # 1. MAJOR version when you make incompatible API changes;
 # 2. MINOR version when you add functionality in a backwards-compatible manner;
 # 3. PATCH version when you make backwards-compatible bug fixes.
-__version__ = "2.0.0"
+__version__ = "2.0.1"
