### 2021-05-10 #24 @agporto
@lwasser  I found a few issues with the examples, but @mluerig has already fixed them.
I am completely satisfied with the responses provided by the author. This is an impressive tool and I hope biologists will rejoice.
