### 2021-05-10 #24 @mluerig
thanks @sdonoughe and @agporto - I have released a new version (2.0.1) that addresses the typos and issues you mentioned.
