+++ phenopype/utils_lowlevel.py
@@ -63,7 +63,26 @@ class _image_viewer:
 
         ## new kwargs
         self.flag_test_mode = kwargs.get("test_mode", False)
-
+        
+        ## create original binary image 
+        self.df_contours = kwargs.get("df_contours", None)
+        if self.df_contours.__class__.__name__ == "DataFrame":
+            self.image_bin = np.zeros_like(image)
+            for index, row in self.df_contours.iterrows():
+                cv2.drawContours(
+                    image=self.image_bin,
+                    contours=[row["coords"]],
+                    contourIdx=0,
+                    thickness=-1,
+                    color=colours["white"],
+                    maxLevel=3,
+                    offset=None,
+                )
+            self.image_bin_copy = copy.deepcopy(self.image_bin)
+            
+            ## convert binary image to colour mask
+            self._bin2colmask(image, self.image_bin)
+            
         ## resize image canvas
         image_width, image_height = image.shape[1], image.shape[0]
         if image_height > window_max_dimension or image_width > window_max_dimension:
@@ -76,6 +95,15 @@ class _image_viewer:
                     ),
                     cv2.INTER_AREA,
                 )
+                if self.df_contours.__class__.__name__ == "DataFrame":
+                    self.colour_mask = cv2.resize(
+                        self.colour_mask_copy,
+                        (
+                            window_max_dimension,
+                            int((window_max_dimension / image_width) * image_height),
+                        ),
+                        cv2.INTER_AREA,
+                    )
             else:
                 canvas = cv2.resize(
                     image,
@@ -85,6 +113,15 @@ class _image_viewer:
                     ),
                     cv2.INTER_AREA,
                 )
+                if self.df_contours.__class__.__name__ == "DataFrame":
+                    self.colour_mask = cv2.resize(
+                        self.colour_mask_copy,
+                        (
+                            window_max_dimension,
+                            int((window_max_dimension / image_width) * image_height),
+                        ),
+                        cv2.INTER_AREA,
+                    )
         else:
             canvas = copy.deepcopy(image)
 
@@ -145,8 +182,12 @@ class _image_viewer:
                 self.colour_current = copy.deepcopy(self.line_colour)
                 self.orig_contours = kwargs.get("contours")
                 self.zoom_flag = True
+                
                 self.colour_left = kwargs.get("colour_left", "white")
-                self.colour_right = kwargs.get("colour_right", "black")
+                self.colour_right = kwargs.get("colour_right", "black")    
+
+            if not kwargs.get("previous"):
+                self._canvasAndColmask()
 
         ## this needs to be decluttered. there should only be one type of list 
         ## (points, not rectangle separately). relevant parameters for the 
@@ -177,6 +218,15 @@ class _image_viewer:
                             self.line_width,
                         )
                     self.point_list = []
+                    self.canvas = self.image_copy[
+                        self.zoom_y1 : self.zoom_y2, self.zoom_x1 : self.zoom_x2,
+                    ]
+                    self.canvas = cv2.resize(
+                        self.canvas,
+                        (self.canvas_width, self.canvas_height),
+                        interpolation=cv2.INTER_LINEAR,
+                    )
+                    self.canvas_copy = copy.deepcopy(self.canvas)
                 elif self.flag_tool == "line" or self.flag_tool == "polyline" or self.flag_tool == "polygon" or self.flag_test_mode == True:
                     for point in self.point_list:
                         cv2.polylines(
@@ -186,25 +236,29 @@ class _image_viewer:
                             self.line_colour,
                             self.line_width,
                         ) 
+                        
+                    self.canvas = self.image_copy[
[…]
