### 2021-05-13 #24 @lwasser
hi @mluerig
To begin - congratulations - 🎉 phenopype is now a part of the pyopensci ecosystem!

Next steps:

----------------------------------------------
🎉  Thank you @jbencook @sdonoughe @agporto for supporting this review process! 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`
- [ ] Add phenopype to the pyOpenSci website. @mluerig , please open a PR to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/packages.yml): to add your package and name to the list of contributors
- [ ] <reviewers-and-maintainers> If you have time and are open to being listed on our website, please add yourselves to [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/contributors.yml) via a pr so we can list you on our website as contributors!

All -- if you have any feedback for us about the review process please feel free to share it here. We are always looking to improve our process and our documentation in the [contributing-guide](https://www.pyopensci.org/contributing-guide). We have also been updating our documentation to improve the process so all feedback is appreciated!
*********
@mluerig  i'm so sorry that this has been a painfully slow process. I took some time to review the package, feedback and  documentation given the tool doesn't work on MAC Os currently (for  understandable  reasons!).  I wanted to be careful about setting precedent in the future.  However, you have that clearly stated in your README file and also requested help / are interested in building out that functionality so we can  consider  that if a tool comes through with similar functionality for MAC.

The past year has  been extremely challenging but starting in August / possibly July this project will have my full attention and  future reviews will NOT be so slow. Thank you again  for  your patience. Please let me know if you have any questions / concerns!  A review  should not take a year so I have that commented noted already :)
