### 2021-05-13 #24 @mluerig
@lwasser: done, done and done! Thanks for wrapping this up so quickly. I will submit this to Methods in Ecology and Evolution in the next few days. Any suggestion for how I should bring up this review there? There are options for rOpenSci, but no pyOpenSci - should I just write this in the cover letter?

Feedback: I would not change anything in the review process; there were templates, instructions and at the start the roles were distributed clearly. My only suggestion for improvement is, as you may have guessed, to work with compulsory deadlines, and a stronger role of the editor to make sure they are kept. Otherwise it gets swept away by the flood of other work we all are involved with - at least I would prefer to have a deadline, simply to mark this in my agenda and get my butt going :-).

In any case, I am extremely grateful for the time you spent on reviewing and editing, and I thank everyone involved. pyOpenSci is a great consortium, and I would submit here again!
