### 2021-05-13 #24 @lwasser
wonderful. Thank you for this feedback.

And I hear you @mluerig i really do. I will note the need for deadlines and followup to ensure a more efficient process. Given I have your ear I have another question for you.
Right  now, you have citation information in your readme. I'm curious if you had thought about a more formal DOI / citation approach through a tool such as zenodo ?

I'm asking because i've been thinking a lot about citations. JOSS provides one by default through their review process.  But Zenodo  can track releases  and provide a citable DOI as well. Do you have any thoughts on this by chance?

i will merge both  pr's shortly to our website. Thank you for doing that so quickly!

https://github.com/pyOpenSci/pyopensci.github.io/pull/62/files
https://github.com/pyOpenSci/pyopensci.github.io/pull/63
