### 2021-05-14 #24 @mluerig
> the need for deadlines and followup
this can really just be a simple post in the issue to remind everyone

@lwasser good point about citations. phenopype already has a zenodo DOI, so I went ahead and added it to the repo

 > I will submit this to Methods in Ecology and Evolution in the next few days. Any suggestion for how I should bring up this review there? There are options for rOpenSci, but no pyOpenSci - should I just write this in the cover letter?
 => do you have any thoughts on this?
