### 2021-05-17 #24 @lwasser
hi @mluerig my apologies i was out of the office Friday.

I think for now it does make sense to include the zenodo reference in the manuscript. And then absolutely include the manuscript reference once you have it. If the article is really methods based (i presume it is given the journal you are submitting to) then i would think once you have a final journal DOI, that citation will be preferred over zenodo.

Yes, rOpenSci has a relationship with them. I have yet to build that relationship but can and would like to once we really get moving on the project in August.  I need to speak with Karthik more about how that relationship works to better understand it. Yes please add this to your cover letter.  And if there is a person that I can reach out to there, please feel free to forward their information to me OR give them my contact information. How does that sound?
