+++ tests/test_03_core_segmentation.py
@@ -15,7 +15,7 @@ def test_blur(project_container):
     assert not (project_container.image_copy==project_container.image).all()
 
 def test_threshold(project_container):
-    pp.segmentation.threshold(project_container, method="adaptive", 
+    pp.segmentation.threshold(project_container, method="adaptive",
                               blocksize=149, constant=10, channel="red")
     unique, counts = np.unique(project_container.image, return_counts=True)
     assert len(unique) == 2
@@ -30,17 +30,7 @@ def test_watershed(project_container):
     thresh = copy.deepcopy(project_container.image)
     water = pp.segmentation.watershed(image, thresh)
     assert not (thresh==water).all()
-
-def test_draw(project_container):
-    test_params = {"flag_test_mode": True,
-                   "line_width": 2,
-                   "line_colour": (0,0,0),
-                   "point_list": [[(1417, 327), (1410, 346)]]}
-    pp.segmentation.draw(project_container, overwrite=flag_overwrite, test_params=test_params)
-    assert project_container.df_drawings.iloc[0]["coords"] == "[(1417, 327), (1410, 346)]"
-    
-    obj_input = project_container
-    
+   
 def test_find_contours(project_container):
     pp.segmentation.find_contours(project_container, min_area=250)
     assert len(project_container.df_contours)>0
