### 2021-05-18 #24 @mluerig
yupp that sounds good - I proceeded with these suggestions!
