### 2021-05-18 #24 @lwasser
Perfect!

Sent from my Verizon, Samsung Galaxy smartphone
Get Outlook for Android<https://aka.ms/AAb9ysg>
