### 2022-09-15 #24 @lwasser
hey 👋  @mluerig   @jbencook  @agporto @sdonoughe  ! I hope that you are all well. I am reaching out here to all reviewers and maintainers about pyOpenSci [now that i am working full time on the project (read more here)](https://www.pyopensci.org/blog/new-executive-director-leah-wasser). We have a survey that we'd like for you to fill out so we can:

[🔗 HERE IS THE SURVEY LINK 🔗 ](https://forms.gle/ZZXcD1hr18tYmY2p6)

1. invite you to our slack channel to participate in our community (if you wish to join - no worries if that is not how you prefer to communicate / participate).
2. Collect information from you about how we can improve our review process and also better serve maintainers.
The survey should take about 10 minutes to complete  depending upon how much you decide to write. This information will help us greatly as we make decisions about how pyOpenSci grows and serves the community. Thank you so much in advance for filling it out.

NOTE: this is different from the form designed for reviewers to sign up to review.
If there are other maintainers for this project, please ping them here and ask them to fill out the survey as well. It is important that we ensure packages are supported long term or sunsetted with sufficient communication to users. Thus we will check in with maintainers annually about maintenance.

Thank you in advance for doing this and supporting pyOpenSci.
