### 2023-03-06 #24 @lwasser
hey there @mluerig i've going through and updating our package catalog and noticed that there is no pyopensci badge on your readme for phenopype! this only came up because the original url here was updated to phenopype/phenopype so i had to search for ya!

i wondered if you were open to adding the reviewed badge to your project readme? Many thanks!

```
[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/24)
```
