### 2023-03-08 #24 @mluerig
Hi Leah – will take care of it as soon as I’m back from vacation!

Cheers
Moritz

From: Leah Wasser ***@***.***>
Sent: Monday, 6 March 2023 23:09
To: pyOpenSci/software-submission ***@***.***>
Cc: Moritz Lürig ***@***.***>; Mention ***@***.***>
Subject: Re: [pyOpenSci/software-submission] Phenopype: a phenotyping pipeline for Python (#24)

hey there @mluerig <https://github.com/mluerig>  i've going through and updating our package catalog and noticed that there is no pyopensci badge on your readme for phenopype! this only came up because the original url here was updated to phenopype/phenopype so i had to search for ya!

i wondered if you were open to adding the reviewed badge to your project readme? Many thanks!

[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/24 <https://tinyurl.com/y22nb8up)%5d(https:/github.com/pyOpenSci/software-review/issues/24> )
