### 2023-04-24 #24 @NickleDave
Hi @mluerig just checking back as we're updating metadata for our pyOpenSci sprint at PyCon -- are you back from vacation and if so could you add our badge when you get a chance?

https://github.com/pyOpenSci/software-submission/issues/24#issuecomment-1457083223

We're happy to make a PR to phenophype adding it to the README if that would help!
